Extract deferred, ssr, view transition, path resolution tests
