+++ packages/router/__tests__/router-test.ts
@@ -45,12 +45,8 @@ import {
   tick,
 } from "./utils/utils";
 
-///////////////////////////////////////////////////////////////////////////////
-//#region Types and Utils
-///////////////////////////////////////////////////////////////////////////////
-
 interface CustomMatchers<R = jest.Expect> {
-  URL(url: string);
+  urlMatch(url: string);
   trackedPromise(data?: any, error?: any, aborted?: boolean): R;
   deferredData(
     done: boolean,
@@ -85,12 +81,6 @@ function initializeTmTest(init?: {
     ...(init?.url ? { initialEntries: [init.url] } : {}),
   });
 }
-
-//#endregion
-
-///////////////////////////////////////////////////////////////////////////////
-//#region Tests
-///////////////////////////////////////////////////////////////////////////////
 
 // Reusable routes for a simple tasks app, for test cases that don't want
 // to create their own more complex routes
@@ -11022,1683 +11012,6 @@ describe("a router", () => {
     });
   });
 
-  describe("deferred data", () => {
-    it("should not track deferred responses on naked objects", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-          },
-          {
-            id: "lazy",
-            path: "lazy",
-            loader: true,
-          },
-        ],
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/lazy");
-
-      let dfd = createDeferred();
-      await A.loaders.lazy.resolve({
-        critical: "1",
-        lazy: dfd.promise,
-      });
-      expect(t.router.state.loaderData).toEqual({
-        lazy: {
-          critical: "1",
-          lazy: expect.any(Promise),
-        },
-      });
-      expect(t.router.state.loaderData.lazy.lazy._tracked).toBeUndefined();
-    });
-
-    it("should support returning deferred responses", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-          },
-          {
-            id: "lazy",
-            path: "lazy",
-            loader: true,
-          },
-        ],
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/lazy");
-
-      let dfd1 = createDeferred();
-      let dfd2 = createDeferred();
-      let dfd3 = createDeferred();
-      dfd1.resolve("Immediate data");
-      await A.loaders.lazy.resolve(
-        defer({
-          critical1: "1",
-          critical2: "2",
-          lazy1: dfd1.promise,
-          lazy2: dfd2.promise,
-          lazy3: dfd3.promise,
-        })
-      );
-      expect(t.router.state.loaderData).toEqual({
-        lazy: {
-          critical1: "1",
-          critical2: "2",
-          lazy1: expect.trackedPromise("Immediate data"),
-          lazy2: expect.trackedPromise(),
-          lazy3: expect.trackedPromise(),
-        },
-      });
-
-      await dfd2.resolve("2");
-      expect(t.router.state.loaderData).toEqual({
-        lazy: {
-          critical1: "1",
-          critical2: "2",
-          lazy1: expect.trackedPromise("Immediate data"),
-          lazy2: expect.trackedPromise("2"),
-          lazy3: expect.trackedPromise(),
-        },
-      });
-
-      await dfd3.resolve("3");
-      expect(t.router.state.loaderData).toEqual({
-        lazy: {
-          critical1: "1",
-          critical2: "2",
-          lazy1: expect.trackedPromise("Immediate data"),
-          lazy2: expect.trackedPromise("2"),
-          lazy3: expect.trackedPromise("3"),
-        },
-      });
-
-      // Should proxy values through
-      let data = t.router.state.loaderData.lazy;
-      await expect(data.lazy1).resolves.toBe("Immediate data");
-      await expect(data.lazy2).resolves.toBe("2");
-      await expect(data.lazy3).resolves.toBe("3");
-    });
-
-    it("should cancel outstanding deferreds on a new navigation", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "lazy",
-            path: "lazy",
-            loader: true,
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/lazy");
-      let dfd1 = createDeferred();
-      let dfd2 = createDeferred();
-      await A.loaders.lazy.resolve(
-        defer({
-          critical1: "1",
-          critical2: "2",
-          lazy1: dfd1.promise,
-          lazy2: dfd2.promise,
-        })
-      );
-
-      // Interrupt pending deferred's from /lazy navigation
-      let navPromise = t.navigate("/");
-
-      // Cancelled promises should reject immediately
-      let data = t.router.state.loaderData.lazy;
-      await expect(data.lazy1).rejects.toBeInstanceOf(AbortedDeferredError);
-      await expect(data.lazy2).rejects.toBeInstanceOf(AbortedDeferredError);
-      await expect(data.lazy1).rejects.toThrowError("Deferred data aborted");
-      await expect(data.lazy2).rejects.toThrowError("Deferred data aborted");
-
-      let B = await navPromise;
-
-      // During navigation - deferreds remain as promises
-      expect(t.router.state.loaderData).toEqual({
-        lazy: {
-          critical1: "1",
-          critical2: "2",
-          lazy1: expect.trackedPromise(null, null, true),
-          lazy2: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      // But they are frozen - no re-paints on resolve/reject!
-      await dfd1.resolve("a");
-      await dfd2.reject(new Error("b"));
-      expect(t.router.state.loaderData).toEqual({
-        lazy: {
-          critical1: "1",
-          critical2: "2",
-          lazy1: expect.trackedPromise(null, null, true),
-          lazy2: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await B.loaders.index.resolve("INDEX*");
-      expect(t.router.state.loaderData).toEqual({
-        index: "INDEX*",
-      });
-    });
-
-    it("should not cancel outstanding deferreds on reused routes", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "root",
-            path: "/",
-            loader: true,
-          },
-          {
-            id: "parent",
-            path: "parent",
-            loader: true,
-            children: [
-              {
-                id: "a",
-                path: "a",
-                loader: true,
-              },
-              {
-                id: "b",
-                path: "b",
-                loader: true,
-              },
-            ],
-          },
-        ],
-        hydrationData: { loaderData: { root: "ROOT" } },
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/parent/a");
-      let parentDfd = createDeferred();
-      await A.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT",
-          lazy: parentDfd.promise,
-        })
-      );
-      let aDfd = createDeferred();
-      await A.loaders.a.resolve(
-        defer({
-          critical: "CRITICAL A",
-          lazy: aDfd.promise,
-        })
-      );
-
-      // Navigate such that we reuse the parent route
-      let B = await t.navigate("/parent/b");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      // This should reflect in loaderData
-      await parentDfd.resolve("LAZY PARENT");
-      // This should not
-      await aDfd.resolve("LAZY A");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise("LAZY PARENT"),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(null, null, true), // No re-paint!
-        },
-      });
-
-      // Complete the navigation
-      await B.loaders.b.resolve("B DATA");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise("LAZY PARENT"),
-        },
-        b: "B DATA",
-      });
-    });
-
-    it("should handle promise rejections", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-          },
-          {
-            id: "lazy",
-            path: "lazy",
-            loader: true,
-          },
-        ],
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/lazy");
-
-      let dfd = createDeferred();
-      await A.loaders.lazy.resolve(
-        defer({
-          critical: "1",
-          lazy: dfd.promise,
-        })
-      );
-
-      await dfd.reject(new Error("Kaboom!"));
-      expect(t.router.state.loaderData).toEqual({
-        lazy: {
-          critical: "1",
-          lazy: expect.trackedPromise(undefined, new Error("Kaboom!")),
-        },
-      });
-
-      // should proxy the error through
-      let data = t.router.state.loaderData.lazy;
-      await expect(data.lazy).rejects.toEqual(new Error("Kaboom!"));
-    });
-
-    it("should cancel all outstanding deferreds on router.revalidate()", async () => {
-      let shouldRevalidateSpy = jest.fn(() => false);
-      let t = setup({
-        routes: [
-          {
-            id: "root",
-            path: "/",
-            loader: true,
-          },
-          {
-            id: "parent",
-            path: "parent",
-            loader: true,
-            shouldRevalidate: shouldRevalidateSpy,
-            children: [
-              {
-                id: "index",
-                index: true,
-                loader: true,
-              },
-            ],
-          },
-        ],
-        hydrationData: { loaderData: { root: "ROOT" } },
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/parent");
-      let parentDfd = createDeferred();
-      await A.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT",
-          lazy: parentDfd.promise,
-        })
-      );
-      let indexDfd = createDeferred();
-      await A.loaders.index.resolve(
-        defer({
-          critical: "CRITICAL INDEX",
-          lazy: indexDfd.promise,
-        })
-      );
-
-      // Trigger a revalidation which should cancel outstanding deferreds
-      let R = await t.revalidate();
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(),
-        },
-        index: {
-          critical: "CRITICAL INDEX",
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      // Neither should reflect in loaderData
-      await parentDfd.resolve("Nope!");
-      await indexDfd.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-        index: {
-          critical: "CRITICAL INDEX",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      // Complete the revalidation
-      let parentDfd2 = createDeferred();
-      await R.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT 2",
-          lazy: parentDfd2.promise,
-        })
-      );
-      let indexDfd2 = createDeferred();
-      await R.loaders.index.resolve(
-        defer({
-          critical: "CRITICAL INDEX 2",
-          lazy: indexDfd2.promise,
-        })
-      );
-
-      // Revalidations await all deferreds, so we're still in a loading
-      // state with the prior loaderData here
-      expect(t.router.state.navigation.state).toBe("idle");
-      expect(t.router.state.revalidation).toBe("loading");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-        index: {
-          critical: "CRITICAL INDEX",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await indexDfd2.resolve("LAZY INDEX 2");
-      // Not done yet!
-      expect(t.router.state.navigation.state).toBe("idle");
-      expect(t.router.state.revalidation).toBe("loading");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-        index: {
-          critical: "CRITICAL INDEX",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await parentDfd2.resolve("LAZY PARENT 2");
-      // Done now that all deferreds have resolved
-      expect(t.router.state.navigation.state).toBe("idle");
-      expect(t.router.state.revalidation).toBe("idle");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT 2",
-          lazy: expect.trackedPromise("LAZY PARENT 2"),
-        },
-        index: {
-          critical: "CRITICAL INDEX 2",
-          lazy: expect.trackedPromise("LAZY INDEX 2"),
-        },
-      });
-
-      expect(shouldRevalidateSpy).not.toHaveBeenCalled();
-    });
-
-    it("cancels correctly on revalidations chains", async () => {
-      let shouldRevalidateSpy = jest.fn(() => false);
-      let t = setup({
-        routes: [
-          {
-            id: "root",
-            path: "/",
-          },
-          {
-            id: "foo",
-            path: "foo",
-            loader: true,
-            shouldRevalidate: shouldRevalidateSpy,
-          },
-        ],
-      });
-
-      let A = await t.navigate("/foo");
-      let dfda = createDeferred();
-      await A.loaders.foo.resolve(
-        defer({
-          critical: "CRITICAL A",
-          lazy: dfda.promise,
-        })
-      );
-      expect(t.router.state.loaderData).toEqual({
-        foo: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      let B = await t.revalidate();
-      let dfdb = createDeferred();
-      // This B data will _never_ make it through - since we will await all of
-      // it and we'll revalidate before it resolves
-      await B.loaders.foo.resolve(
-        defer({
-          critical: "CRITICAL B",
-          lazy: dfdb.promise,
-        })
-      );
-      // The initial revalidation cancelled the navigation deferred
-      await dfda.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({
-        foo: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      let C = await t.revalidate();
-      let dfdc = createDeferred();
-      await C.loaders.foo.resolve(
-        defer({
-          critical: "CRITICAL C",
-          lazy: dfdc.promise,
-        })
-      );
-      // The second revalidation should have cancelled the first revalidation
-      // deferred
-      await dfdb.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({
-        foo: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      // Resolve the final revalidation which should make it into loaderData
-      await dfdc.resolve("Yep!");
-      expect(t.router.state.loaderData).toEqual({
-        foo: {
-          critical: "CRITICAL C",
-          lazy: expect.trackedPromise("Yep!"),
-        },
-      });
-
-      expect(shouldRevalidateSpy).not.toHaveBeenCalled();
-    });
-
-    it("cancels correctly on revalidations interrupted by navigations", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "root",
-            path: "/",
-          },
-          {
-            id: "foo",
-            path: "foo",
-            loader: true,
-          },
-          {
-            id: "bar",
-            path: "bar",
-            loader: true,
-          },
-        ],
-      });
-
-      let A = await t.navigate("/foo");
-      let dfda = createDeferred();
-      await A.loaders.foo.resolve(
-        defer({
-          critical: "CRITICAL A",
-          lazy: dfda.promise,
-        })
-      );
-      await dfda.resolve("LAZY A");
-      expect(t.router.state.loaderData).toEqual({
-        foo: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise("LAZY A"),
-        },
-      });
-
-      let B = await t.revalidate();
-      let dfdb = createDeferred();
-      await B.loaders.foo.resolve(
-        defer({
-          critical: "CRITICAL B",
-          lazy: dfdb.promise,
-        })
-      );
-      // B not reflected because its got existing loaderData
-      expect(t.router.state.loaderData).toEqual({
-        foo: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise("LAZY A"),
-        },
-      });
-
-      let C = await t.navigate("/bar");
-      let dfdc = createDeferred();
-      await C.loaders.bar.resolve(
-        defer({
-          critical: "CRITICAL C",
-          lazy: dfdc.promise,
-        })
-      );
-      // The second revalidation should have cancelled the first revalidation
-      // deferred
-      await dfdb.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({
-        bar: {
-          critical: "CRITICAL C",
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      await dfdc.resolve("Yep!");
-      expect(t.router.state.loaderData).toEqual({
-        bar: {
-          critical: "CRITICAL C",
-          lazy: expect.trackedPromise("Yep!"),
-        },
-      });
-    });
-
-    it("cancels pending deferreds on 404 navigations", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "lazy",
-            path: "lazy",
-            loader: true,
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/lazy");
-      let dfd = createDeferred();
-      await A.loaders.lazy.resolve(
-        defer({
-          critical: "CRITICAL",
-          lazy: dfd.promise,
-        })
-      );
-
-      await t.navigate("/not-found");
-      // Navigation completes immediately and deferreds are cancelled
-      expect(t.router.state.loaderData).toEqual({});
-
-      // Resolution doesn't do anything
-      await dfd.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({});
-    });
-
-    it("cancels pending deferreds on errored GET submissions (w/ reused routes)", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "parent",
-            path: "parent",
-            loader: true,
-            hasErrorBoundary: true,
-            children: [
-              {
-                id: "a",
-                path: "a",
-                loader: true,
-              },
-              {
-                id: "b",
-                path: "b",
-                loader: true,
-              },
-            ],
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      // Navigate to /parent/a and kick off a deferred's for both
-      let A = await t.navigate("/parent/a");
-      let parentDfd = createDeferred();
-      await A.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT",
-          lazy: parentDfd.promise,
-        })
-      );
-      let aDfd = createDeferred();
-      await A.loaders.a.resolve(
-        defer({
-          critical: "CRITICAL A",
-          lazy: aDfd.promise,
-        })
-      );
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      // Perform an invalid navigation to /parent/b which will be handled
-      // using parent's error boundary.  Parent's deferred should be left alone
-      // while A's should be cancelled since they will no longer be rendered
-      let B = await t.navigate("/parent/b");
-      await B.loaders.b.reject(
-        new Response("broken", { status: 400, statusText: "Bad Request" })
-      );
-
-      // Navigation completes immediately with an error at the boundary
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(),
-        },
-      });
-      expect(t.router.state.errors).toEqual({
-        parent: new ErrorResponseImpl(400, "Bad Request", "broken", false),
-      });
-
-      await parentDfd.resolve("Yep!");
-      await aDfd.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise("Yep!"),
-        },
-      });
-    });
-
-    it("cancels pending deferreds on errored GET submissions (w/o reused routes)", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "a",
-            path: "a",
-            loader: true,
-            children: [
-              {
-                id: "aChild",
-                path: "child",
-                loader: true,
-              },
-            ],
-          },
-          {
-            id: "b",
-            path: "b",
-            loader: true,
-            children: [
-              {
-                id: "bChild",
-                path: "child",
-                loader: true,
-                hasErrorBoundary: true,
-              },
-            ],
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      // Navigate to /parent/a and kick off deferred's for both
-      let A = await t.navigate("/a/child");
-      let aDfd = createDeferred();
-      await A.loaders.a.resolve(
-        defer({
-          critical: "CRITICAL A",
-          lazy: aDfd.promise,
-        })
-      );
-      let aChildDfd = createDeferred();
-      await A.loaders.aChild.resolve(
-        defer({
-          critical: "CRITICAL A CHILD",
-          lazy: aChildDfd.promise,
-        })
-      );
-      expect(t.router.state.loaderData).toEqual({
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(),
-        },
-        aChild: {
-          critical: "CRITICAL A CHILD",
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      // Perform an invalid navigation to /b/child which should cancel all
-      // pending deferred's since nothing is reused.  It should not call bChild's
-      // loader since it's below the boundary but should call b's loader.
-      let B = await t.navigate("/b/child");
-
-      await B.loaders.bChild.reject(
-        new Response("broken", { status: 400, statusText: "Bad Request" })
-      );
-
-      // Both should be cancelled
-      await aDfd.resolve("Nope!");
-      await aChildDfd.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-        aChild: {
-          critical: "CRITICAL A CHILD",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await B.loaders.b.resolve("B LOADER");
-      expect(t.router.state.loaderData).toEqual({
-        b: "B LOADER",
-      });
-      expect(t.router.state.errors).toEqual({
-        bChild: new ErrorResponseImpl(400, "Bad Request", "broken", false),
-      });
-    });
-
-    it("does not cancel pending deferreds on hash change only navigations", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "lazy",
-            path: "lazy",
-            loader: true,
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/lazy");
-      let dfd = createDeferred();
-      await A.loaders.lazy.resolve(
-        defer({
-          critical: "CRITICAL",
-          lazy: dfd.promise,
-        })
-      );
-
-      await t.navigate("/lazy#hash");
-      expect(t.router.state.loaderData).toEqual({
-        lazy: {
-          critical: "CRITICAL",
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      await dfd.resolve("Yep!");
-      expect(t.router.state.loaderData).toEqual({
-        lazy: {
-          critical: "CRITICAL",
-          lazy: expect.trackedPromise("Yep!"),
-        },
-      });
-    });
-
-    it("cancels pending deferreds on action submissions", async () => {
-      let shouldRevalidateSpy = jest.fn(() => false);
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "parent",
-            path: "parent",
-            loader: true,
-            shouldRevalidate: shouldRevalidateSpy,
-            children: [
-              {
-                id: "a",
-                path: "a",
-                loader: true,
-              },
-              {
-                id: "b",
-                path: "b",
-                action: true,
-              },
-            ],
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/parent/a");
-      let parentDfd = createDeferred();
-      await A.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT",
-          lazy: parentDfd.promise,
-        })
-      );
-      let aDfd = createDeferred();
-      await A.loaders.a.resolve(
-        defer({
-          critical: "CRITICAL A",
-          lazy: aDfd.promise,
-        })
-      );
-
-      // Action submission causes all to be cancelled, even reused ones, and
-      // ignores shouldRevalidate since the cancelled active deferred means we
-      // are missing data
-      let B = await t.navigate("/parent/b", {
-        formMethod: "post",
-        formData: createFormData({ key: "value" }),
-      });
-      await parentDfd.resolve("Nope!");
-      await aDfd.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await B.actions.b.resolve("ACTION");
-      let parentDfd2 = createDeferred();
-      await B.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT 2",
-          lazy: parentDfd2.promise,
-        })
-      );
-      expect(t.router.state.actionData).toEqual({
-        b: "ACTION",
-      });
-      // Since we still have outstanding deferreds on the revalidation, we're
-      // still in the loading state and showing the old data
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await parentDfd2.resolve("Yep!");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT 2",
-          lazy: expect.trackedPromise("Yep!"),
-        },
-      });
-
-      expect(shouldRevalidateSpy).not.toHaveBeenCalled();
-    });
-
-    it("does not put resolved deferred's back into a loading state during revalidation", async () => {
-      let shouldRevalidateSpy = jest.fn(() => false);
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "parent",
-            path: "parent",
-            loader: true,
-            shouldRevalidate: shouldRevalidateSpy,
-            children: [
-              {
-                id: "a",
-                path: "a",
-                loader: true,
-              },
-              {
-                id: "b",
-                path: "b",
-                action: true,
-                loader: true,
-              },
-            ],
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      // Route to /parent/a and return and resolve deferred's for both
-      let A = await t.navigate("/parent/a");
-      let parentDfd1 = createDeferred();
-      let parentDfd2 = createDeferred();
-      await A.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT",
-          lazy1: parentDfd1.promise,
-          lazy2: parentDfd2.promise,
-        })
-      );
-      let aDfd1 = createDeferred();
-      let aDfd2 = createDeferred();
-      await A.loaders.a.resolve(
-        defer({
-          critical: "CRITICAL A",
-          lazy1: aDfd1.promise,
-          lazy2: aDfd2.promise,
-        })
-      );
-
-      // Resolve one of the deferred for each prior to the action submission
-      await parentDfd1.resolve("LAZY PARENT 1");
-      await aDfd1.resolve("LAZY A 1");
-
-      // Action submission causes all to be cancelled, even reused ones, and
-      // ignores shouldRevalidate since the cancelled active deferred means we
-      // are missing data
-      let B = await t.navigate("/parent/b", {
-        formMethod: "post",
-        formData: createFormData({ key: "value" }),
-      });
-      await parentDfd2.resolve("Nope!");
-      await aDfd2.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy1: expect.trackedPromise("LAZY PARENT 1"),
-          lazy2: expect.trackedPromise(null, null, true),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy1: expect.trackedPromise("LAZY A 1"),
-          lazy2: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await B.actions.b.resolve("ACTION");
-      let parentDfd1Revalidation = createDeferred();
-      let parentDfd2Revalidation = createDeferred();
-      await B.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT*",
-          lazy1: parentDfd1Revalidation.promise,
-          lazy2: parentDfd2Revalidation.promise,
-        })
-      );
-      await B.loaders.b.resolve("B");
-
-      // At this point, we resolved the action and the loaders - however the
-      // parent loader returned a deferred so we stay in the "loading" state
-      // until everything resolves
-      expect(t.router.state.navigation.state).toBe("loading");
-      expect(t.router.state.actionData).toEqual({
-        b: "ACTION",
-      });
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy1: expect.trackedPromise("LAZY PARENT 1"),
-          lazy2: expect.trackedPromise(null, null, true),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy1: expect.trackedPromise("LAZY A 1"),
-          lazy2: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      // Resolve the first deferred - should not complete the navigation yet
-      await parentDfd1Revalidation.resolve("LAZY PARENT 1*");
-      expect(t.router.state.navigation.state).toBe("loading");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy1: expect.trackedPromise("LAZY PARENT 1"),
-          lazy2: expect.trackedPromise(null, null, true),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy1: expect.trackedPromise("LAZY A 1"),
-          lazy2: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await parentDfd2Revalidation.resolve("LAZY PARENT 2*");
-      expect(t.router.state.navigation.state).toBe("idle");
-      expect(t.router.state.actionData).toEqual({
-        b: "ACTION",
-      });
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT*",
-          lazy1: expect.trackedPromise("LAZY PARENT 1*"),
-          lazy2: expect.trackedPromise("LAZY PARENT 2*"),
-        },
-        b: "B",
-      });
-
-      expect(shouldRevalidateSpy).not.toHaveBeenCalled();
-    });
-
-    it("triggers fallbacks on new dynamic route instances", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "invoice",
-            path: "invoices/:id",
-            loader: true,
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/invoices/1");
-      let dfd1 = createDeferred();
-      await A.loaders.invoice.resolve(defer({ lazy: dfd1.promise }));
-      expect(t.router.state.loaderData).toEqual({
-        invoice: {
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      await dfd1.resolve("DATA 1");
-      expect(t.router.state.loaderData).toEqual({
-        invoice: {
-          lazy: expect.trackedPromise("DATA 1"),
-        },
-      });
-
-      // Goes back into a loading state since this is a new instance of the
-      // invoice route
-      let B = await t.navigate("/invoices/2");
-      let dfd2 = createDeferred();
-      await B.loaders.invoice.resolve(defer({ lazy: dfd2.promise }));
-      expect(t.router.state.loaderData).toEqual({
-        invoice: {
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      await dfd2.resolve("DATA 2");
-      expect(t.router.state.loaderData).toEqual({
-        invoice: {
-          lazy: expect.trackedPromise("DATA 2"),
-        },
-      });
-    });
-
-    it("triggers fallbacks on new splat route instances", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "invoices",
-            path: "invoices",
-            children: [
-              {
-                id: "invoice",
-                path: "*",
-                loader: true,
-              },
-            ],
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/invoices/1");
-      let dfd1 = createDeferred();
-      await A.loaders.invoice.resolve(defer({ lazy: dfd1.promise }));
-      expect(t.router.state.loaderData).toEqual({
-        invoice: {
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      await dfd1.resolve("DATA 1");
-      expect(t.router.state.loaderData).toEqual({
-        invoice: {
-          lazy: expect.trackedPromise("DATA 1"),
-        },
-      });
-
-      // Goes back into a loading state since this is a new instance of the
-      // invoice route
-      let B = await t.navigate("/invoices/2");
-      let dfd2 = createDeferred();
-      await B.loaders.invoice.resolve(defer({ lazy: dfd2.promise }));
-      expect(t.router.state.loaderData).toEqual({
-        invoice: {
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      await dfd2.resolve("DATA 2");
-      expect(t.router.state.loaderData).toEqual({
-        invoice: {
-          lazy: expect.trackedPromise("DATA 2"),
-        },
-      });
-    });
-
-    it("cancels awaited reused deferreds on subsequent navigations", async () => {
-      let shouldRevalidateSpy = jest.fn(() => false);
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "parent",
-            path: "parent",
-            loader: true,
-            shouldRevalidate: shouldRevalidateSpy,
-            children: [
-              {
-                id: "a",
-                path: "a",
-                loader: true,
-              },
-              {
-                id: "b",
-                path: "b",
-                action: true,
-                loader: true,
-              },
-            ],
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      // Route to /parent/a and return and resolve deferred's for both
-      let A = await t.navigate("/parent/a");
-      let parentDfd = createDeferred(); // Never resolves in this test
-      await A.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT",
-          lazy: parentDfd.promise,
-        })
-      );
-      let aDfd = createDeferred();
-      await A.loaders.a.resolve(
-        defer({
-          critical: "CRITICAL A",
-          lazy: aDfd.promise,
-        })
-      );
-
-      // Action submission to cancel deferreds
-      let B = await t.navigate("/parent/b", {
-        formMethod: "post",
-        formData: createFormData({ key: "value" }),
-      });
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(),
-        },
-      });
-
-      await B.actions.b.resolve("ACTION");
-      let parentDfd2 = createDeferred(); // Never resolves in this test
-      await B.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT*",
-          lazy: parentDfd2.promise,
-        })
-      );
-      await B.loaders.b.resolve("B");
-
-      // Still in loading state due to revalidation deferred
-      expect(t.router.state.navigation.state).toBe("loading");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      // Navigate elsewhere - should cancel/abort revalidation deferreds
-      let C = await t.navigate("/");
-      await C.loaders.index.resolve("INDEX*");
-      expect(t.router.state.navigation.state).toBe("idle");
-      expect(t.router.state.actionData).toEqual(null);
-      expect(t.router.state.loaderData).toEqual({
-        index: "INDEX*",
-      });
-    });
-
-    it("does not support deferred data on fetcher loads", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-          },
-          {
-            id: "fetch",
-            path: "fetch",
-            loader: true,
-          },
-        ],
-        initialEntries: ["/"],
-      });
-
-      let key = "key";
-      let A = await t.fetch("/fetch", key);
-
-      // deferred in a fetcher awaits all data in the loading state
-      let dfd = createDeferred();
-      await A.loaders.fetch.resolve(
-        defer({
-          critical: "1",
-          lazy: dfd.promise,
-        })
-      );
-      expect(t.router.state.fetchers.get(key)).toMatchObject({
-        state: "loading",
-        data: undefined,
-      });
-
-      await dfd.resolve("2");
-      expect(t.router.state.fetchers.get(key)).toMatchObject({
-        state: "idle",
-        data: {
-          critical: "1",
-          lazy: "2",
-        },
-      });
-
-      // Trigger a revalidation for the same fetcher
-      let B = await t.revalidate("fetch", "fetch");
-      expect(t.router.state.revalidation).toBe("loading");
-      let dfd2 = createDeferred();
-      await B.loaders.fetch.resolve(
-        defer({
-          critical: "3",
-          lazy: dfd2.promise,
-        })
-      );
-      expect(t.router.state.fetchers.get(key)).toMatchObject({
-        state: "idle",
-        data: {
-          critical: "1",
-          lazy: "2",
-        },
-      });
-
-      await dfd2.resolve("4");
-      expect(t.router.state.fetchers.get(key)).toMatchObject({
-        state: "idle",
-        data: {
-          critical: "3",
-          lazy: "4",
-        },
-      });
-    });
-
-    it("triggers error boundaries if fetcher deferred data rejects", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-          },
-          {
-            id: "fetch",
-            path: "fetch",
-            loader: true,
-          },
-        ],
-        initialEntries: ["/"],
-      });
-
-      let key = "key";
-      let A = await t.fetch("/fetch", key);
-
-      let dfd = createDeferred();
-      await A.loaders.fetch.resolve(
-        defer({
-          critical: "1",
-          lazy: dfd.promise,
-        })
-      );
-      await dfd.reject(new Error("Kaboom!"));
-      expect(t.router.state.errors).toMatchObject({
-        index: new Error("Kaboom!"),
-      });
-      expect(t.router.state.fetchers.get(key)).toBeUndefined();
-    });
-
-    it("cancels pending deferreds on fetcher reloads", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-          },
-          {
-            id: "fetch",
-            path: "fetch",
-            loader: true,
-          },
-        ],
-        initialEntries: ["/"],
-      });
-
-      let key = "key";
-      let A = await t.fetch("/fetch", key);
-
-      // deferred in a fetcher awaits all data in the loading state
-      let dfd1 = createDeferred();
-      let loaderPromise1 = A.loaders.fetch.resolve(
-        defer({
-          critical: "1",
-          lazy: dfd1.promise,
-        })
-      );
-      expect(t.router.state.fetchers.get(key)).toMatchObject({
-        state: "loading",
-        data: undefined,
-      });
-
-      // Fetch again
-      let B = await t.fetch("/fetch", key);
-
-      let dfd2 = createDeferred();
-      let loaderPromise2 = B.loaders.fetch.resolve(
-        defer({
-          critical: "3",
-          lazy: dfd2.promise,
-        })
-      );
-      expect(t.router.state.fetchers.get(key)).toMatchObject({
-        state: "loading",
-        data: undefined,
-      });
-
-      // Resolving the second finishes us up
-      await dfd1.resolve("2");
-      await dfd2.resolve("4");
-      await loaderPromise1;
-      await loaderPromise2;
-      expect(t.router.state.fetchers.get(key)).toMatchObject({
-        state: "idle",
-        data: {
-          critical: "3",
-          lazy: "4",
-        },
-      });
-    });
-
-    it("cancels pending deferreds on fetcher action submissions", async () => {
-      let t = setup({
-        routes: [
-          {
-            id: "index",
-            index: true,
-            loader: true,
-          },
-          {
-            id: "parent",
-            path: "parent",
-            loader: true,
-            shouldRevalidate: () => false,
-            children: [
-              {
-                id: "a",
-                path: "a",
-                loader: true,
-              },
-              {
-                id: "b",
-                path: "b",
-                action: true,
-              },
-            ],
-          },
-        ],
-        hydrationData: { loaderData: { index: "INDEX" } },
-        initialEntries: ["/"],
-      });
-
-      let A = await t.navigate("/parent/a");
-      let parentDfd = createDeferred();
-      await A.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT",
-          lazy: parentDfd.promise,
-        })
-      );
-      let aDfd = createDeferred();
-      await A.loaders.a.resolve(
-        defer({
-          critical: "CRITICAL A",
-          lazy: aDfd.promise,
-        })
-      );
-
-      // Fetcher action submission causes all to be cancelled and
-      // ignores shouldRevalidate since the cancelled active deferred means we
-      // are missing data
-      let key = "key";
-      let B = await t.fetch("/parent/b", key, {
-        formMethod: "post",
-        formData: createFormData({ key: "value" }),
-      });
-      await parentDfd.resolve("Nope!");
-      await aDfd.resolve("Nope!");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await B.actions.b.resolve("ACTION");
-      expect(t.router.state.fetchers.get(key)).toMatchObject({
-        state: "loading",
-        data: "ACTION",
-      });
-
-      await B.actions.b.resolve("ACTION");
-      let parentDfd2 = createDeferred();
-      await B.loaders.parent.resolve(
-        defer({
-          critical: "CRITICAL PARENT 2",
-          lazy: parentDfd2.promise,
-        })
-      );
-      let aDfd2 = createDeferred();
-      await B.loaders.a.resolve(
-        defer({
-          critical: "CRITICAL A 2",
-          lazy: aDfd2.promise,
-        })
-      );
-
-      // Still showing old data while we wait on revalidation deferreds to
-      // complete
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-        a: {
-          critical: "CRITICAL A",
-          lazy: expect.trackedPromise(null, null, true),
-        },
-      });
-
-      await parentDfd2.resolve("Yep!");
-      await aDfd2.resolve("Yep!");
-      expect(t.router.state.loaderData).toEqual({
-        parent: {
-          critical: "CRITICAL PARENT 2",
-          lazy: expect.trackedPromise("Yep!"),
-        },
-        a: {
-          critical: "CRITICAL A 2",
-          lazy: expect.trackedPromise("Yep!"),
-        },
-      });
-      expect(t.router.state.fetchers.get(key)).toMatchObject({
-        state: "idle",
-        data: "ACTION",
-      });
-    });
-
-    it("differentiates between navigation and fetcher deferreds on cancellations", async () => {
-      let dfds: Array<ReturnType<typeof createDeferred>> = [];
-      let signals: Array<AbortSignal> = [];
-      let router = createRouter({
-        history: createMemoryHistory({ initialEntries: ["/"] }),
-        routes: [
-          {
-            id: "root",
-            path: "/",
-            loader: ({ request }) => {
-              let dfd = createDeferred();
-              dfds.push(dfd);
-              signals.push(request.signal);
-              return defer({ value: dfd.promise });
-            },
-          },
-        ],
-        hydrationData: {
-          loaderData: {
-            root: { value: -1 },
-          },
-        },
-      });
-
-      // navigate to root, kicking off a reload of the root loader
-      let key = "key";
-      router.navigate("/");
-      router.fetch(key, "root", "/");
-      await tick();
-      expect(router.state.navigation.state).toBe("loading");
-      expect(router.state.loaderData).toEqual({
-        root: { value: -1 },
-      });
-      expect(router.state.fetchers.get(key)).toMatchObject({
-        state: "loading",
-        data: undefined,
-      });
-
-      // Interrupt with a revalidation
-      router.revalidate();
-
-      // Original deferreds should do nothing on resolution
-      dfds[0].resolve(0);
-      dfds[1].resolve(1);
-      await tick();
-      expect(router.state.navigation.state).toBe("loading");
-      expect(router.state.loaderData).toEqual({
-        root: { value: -1 },
-      });
-      expect(router.state.fetchers.get(key)).toMatchObject({
-        state: "loading",
-        data: undefined,
-      });
-
-      // New deferreds should complete the revalidation
-      dfds[2].resolve(2);
-      dfds[3].resolve(3);
-      await tick();
-      expect(router.state.navigation.state).toBe("idle");
-      expect(router.state.loaderData).toEqual({
-        root: { value: expect.trackedPromise(2) },
-      });
-      expect(router.state.fetchers.get(key)).toMatchObject({
-        state: "idle",
-        data: { value: 3 },
-      });
-
-      // Assert that both the route loader and fetcher loader were aborted
-      expect(signals[0].aborted).toBe(true); // initial route
-      expect(signals[1].aborted).toBe(true); // initial fetcher
-      expect(signals[2].aborted).toBe(false); // revalidating route
-      expect(signals[3].aborted).toBe(false); // revalidating fetcher
-
-      expect(router._internalActiveDeferreds.size).toBe(0);
-      expect(router._internalFetchControllers.size).toBe(0);
-      router.dispose();
-    });
-  });
-
   describe("lazily loaded route modules", () => {
     const LAZY_ROUTES: TestRouteObject[] = [
       {
@@ -13835,2187 +12148,6 @@ describe("a router", () => {
     });
   });
 
-  describe("ssr", () => {
-    const SSR_ROUTES = [
-      {
-        id: "index",
-        path: "/",
-        loader: () => "INDEX LOADER",
-      },
-      {
-        id: "parent",
-        path: "/parent",
-        loader: () => "PARENT LOADER",
-        action: () => "PARENT ACTION",
-        children: [
-          {
-            id: "parentIndex",
-            index: true,
-            loader: () => "PARENT INDEX LOADER",
-            action: () => "PARENT INDEX ACTION",
-          },
-          {
-            id: "child",
-            path: "child",
-            loader: () => "CHILD LOADER",
-            action: () => "CHILD ACTION",
-          },
-          {
-            id: "json",
-            path: "json",
-            loader: () => json({ type: "loader" }),
-            action: () => json({ type: "action" }),
-          },
-          {
-            id: "deferred",
-            path: "deferred",
-            loader: ({ request }) => {
-              if (new URL(request.url).searchParams.has("reject")) {
-                return defer({
-                  critical: "loader",
-                  lazy: new Promise((_, r) =>
-                    setTimeout(() => r(new Error("broken!")), 10)
-                  ),
-                });
-              }
-              if (new URL(request.url).searchParams.has("undefined")) {
-                return defer({
-                  critical: "loader",
-                  lazy: new Promise((r) => setTimeout(() => r(undefined), 10)),
-                });
-              }
-              if (new URL(request.url).searchParams.has("status")) {
-                return defer(
-                  {
-                    critical: "loader",
-                    lazy: new Promise((r) => setTimeout(() => r("lazy"), 10)),
-                  },
-                  { status: 201, headers: { "X-Custom": "yes" } }
-                );
-              }
-              return defer({
-                critical: "loader",
-                lazy: new Promise((r) => setTimeout(() => r("lazy"), 10)),
-              });
-            },
-            action: () =>
-              defer({
-                critical: "critical",
-                lazy: new Promise((r) => setTimeout(() => r("lazy"), 10)),
-              }),
-          },
-          {
-            id: "error",
-            path: "error",
-            loader: () => Promise.reject("ERROR LOADER ERROR"),
-            action: () => Promise.reject("ERROR ACTION ERROR"),
-          },
-          {
-            id: "errorBoundary",
-            path: "error-boundary",
-            hasErrorBoundary: true,
-            loader: () => Promise.reject("ERROR BOUNDARY LOADER ERROR"),
-            action: () => Promise.reject("ERROR BOUNDARY ACTION ERROR"),
-          },
-        ],
-      },
-      {
-        id: "redirect",
-        path: "/redirect",
-        loader: () => redirect("/"),
-      },
-    ];
-
-    // Regardless of if the URL is internal or external - all absolute URL
-    // responses should return untouched during SSR so the browser can handle
-    // them
-    let ABSOLUTE_URLS = [
-      "http://localhost/",
-      "https://localhost/about",
-      "http://remix.run/blog",
-      "https://remix.run/blog",
-      "//remix.run/blog",
-      "app://whatever",
-      "mailto:hello@remix.run",
-      "web+remix:whatever",
-    ];
-
-    describe("document requests", () => {
-      it("should support document load navigations", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context = await query(createRequest("/parent/child"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            parent: "PARENT LOADER",
-            child: "CHILD LOADER",
-          },
-          errors: null,
-          location: { pathname: "/parent/child" },
-          matches: [{ route: { id: "parent" } }, { route: { id: "child" } }],
-        });
-      });
-
-      it("should support document load navigations with HEAD requests", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context = await query(
-          createRequest("/parent/child", { method: "HEAD" })
-        );
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            parent: "PARENT LOADER",
-            child: "CHILD LOADER",
-          },
-          errors: null,
-          location: { pathname: "/parent/child" },
-          matches: [{ route: { id: "parent" } }, { route: { id: "child" } }],
-        });
-      });
-
-      it("should support document load navigations with a basename", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES, { basename: "/base" });
-        let context = await query(createRequest("/base/parent/child"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            parent: "PARENT LOADER",
-            child: "CHILD LOADER",
-          },
-          errors: null,
-          location: { pathname: "/base/parent/child" },
-          matches: [{ route: { id: "parent" } }, { route: { id: "child" } }],
-        });
-      });
-
-      it("should fill in null loaderData values for routes without loaders", async () => {
-        let { query } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-            children: [
-              {
-                id: "none",
-                path: "none",
-              },
-              {
-                id: "a",
-                path: "a",
-                loader: () => "A",
-                children: [
-                  {
-                    id: "b",
-                    path: "b",
-                  },
-                ],
-              },
-            ],
-          },
-        ]);
-
-        // No loaders at all
-        let context = await query(createRequest("/none"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            root: null,
-            none: null,
-          },
-          errors: null,
-          location: { pathname: "/none" },
-        });
-
-        // Mix of loaders and no loaders
-        context = await query(createRequest("/a/b"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            root: null,
-            a: "A",
-            b: null,
-          },
-          errors: null,
-          location: { pathname: "/a/b" },
-        });
-      });
-
-      it("should support document load navigations returning responses", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context = await query(createRequest("/parent/json"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            parent: "PARENT LOADER",
-            json: { type: "loader" },
-          },
-          errors: null,
-          matches: [{ route: { id: "parent" } }, { route: { id: "json" } }],
-        });
-      });
-
-      it("should support document load navigations returning deferred", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context = await query(createRequest("/parent/deferred"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            parent: "PARENT LOADER",
-            deferred: {
-              critical: "loader",
-              lazy: expect.trackedPromise(),
-            },
-          },
-          activeDeferreds: {
-            deferred: expect.deferredData(false),
-          },
-          errors: null,
-          location: { pathname: "/parent/deferred" },
-          matches: [{ route: { id: "parent" } }, { route: { id: "deferred" } }],
-        });
-
-        await new Promise((r) => setTimeout(r, 10));
-
-        expect(context).toMatchObject({
-          loaderData: {
-            deferred: {
-              lazy: expect.trackedPromise("lazy"),
-            },
-          },
-          activeDeferreds: {
-            deferred: expect.deferredData(true),
-          },
-        });
-      });
-
-      it("should support document submit navigations", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context = await query(createSubmitRequest("/parent/child"));
-        expect(context).toMatchObject({
-          actionData: {
-            child: "CHILD ACTION",
-          },
-          loaderData: {
-            parent: "PARENT LOADER",
-            child: "CHILD LOADER",
-          },
-          errors: null,
-          location: { pathname: "/parent/child" },
-          matches: [{ route: { id: "parent" } }, { route: { id: "child" } }],
-        });
-      });
-
-      it("should support alternative submission methods", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context;
-
-        let expected = {
-          actionData: {
-            child: "CHILD ACTION",
-          },
-          loaderData: {
-            parent: "PARENT LOADER",
-            child: "CHILD LOADER",
-          },
-          errors: null,
-          location: { pathname: "/parent/child" },
-          matches: [{ route: { id: "parent" } }, { route: { id: "child" } }],
-        };
-
-        context = await query(
-          createSubmitRequest("/parent/child", { method: "PUT" })
-        );
-        expect(context).toMatchObject(expected);
-
-        context = await query(
-          createSubmitRequest("/parent/child", { method: "PATCH" })
-        );
-        expect(context).toMatchObject(expected);
-
-        context = await query(
-          createSubmitRequest("/parent/child", { method: "DELETE" })
-        );
-        expect(context).toMatchObject(expected);
-      });
-
-      it("should support document submit navigations returning responses", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context = await query(createSubmitRequest("/parent/json"));
-        expect(context).toMatchObject({
-          actionData: {
-            json: { type: "action" },
-          },
-          loaderData: {
-            parent: "PARENT LOADER",
-            json: { type: "loader" },
-          },
-          errors: null,
-          matches: [{ route: { id: "parent" } }, { route: { id: "json" } }],
-        });
-      });
-
-      it("should support document submit navigations to layout routes", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context = await query(createSubmitRequest("/parent"));
-        expect(context).toMatchObject({
-          actionData: {
-            parent: "PARENT ACTION",
-          },
-          loaderData: {
-            parent: "PARENT LOADER",
-            parentIndex: "PARENT INDEX LOADER",
-          },
-          errors: null,
-          matches: [
-            { route: { id: "parent" } },
-            { route: { id: "parentIndex" } },
-          ],
-        });
-      });
-
-      it("should support document submit navigations to index routes", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context = await query(createSubmitRequest("/parent?index"));
-        expect(context).toMatchObject({
-          actionData: {
-            parentIndex: "PARENT INDEX ACTION",
-          },
-          loaderData: {
-            parent: "PARENT LOADER",
-            parentIndex: "PARENT INDEX LOADER",
-          },
-          errors: null,
-          matches: [
-            { route: { id: "parent" } },
-            { route: { id: "parentIndex" } },
-          ],
-        });
-      });
-
-      it("should handle redirect Responses", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let response = await query(createRequest("/redirect"));
-        expect(response instanceof Response).toBe(true);
-        expect((response as Response).status).toBe(302);
-        expect((response as Response).headers.get("Location")).toBe("/");
-      });
-
-      it("should handle relative redirect responses (loader)", async () => {
-        let { query } = createStaticHandler([
-          {
-            path: "/",
-            children: [
-              {
-                path: "parent",
-                children: [
-                  {
-                    path: "child",
-                    loader: () => redirect(".."),
-                  },
-                ],
-              },
-            ],
-          },
-        ]);
-        let response = await query(createRequest("/parent/child"));
-        expect(response instanceof Response).toBe(true);
-        expect((response as Response).status).toBe(302);
-        expect((response as Response).headers.get("Location")).toBe("/parent");
-      });
-
-      it("should handle relative redirect responses (action)", async () => {
-        let { query } = createStaticHandler([
-          {
-            path: "/",
-            children: [
-              {
-                path: "parent",
-                children: [
-                  {
-                    path: "child",
-                    action: () => redirect(".."),
-                  },
-                ],
-              },
-            ],
-          },
-        ]);
-        let response = await query(createSubmitRequest("/parent/child"));
-        expect(response instanceof Response).toBe(true);
-        expect((response as Response).status).toBe(302);
-        expect((response as Response).headers.get("Location")).toBe("/parent");
-      });
-
-      it("should handle absolute redirect Responses", async () => {
-        for (let url of ABSOLUTE_URLS) {
-          let handler = createStaticHandler([
-            {
-              path: "/",
-              loader: () => redirect(url),
-            },
-          ]);
-          let response = await handler.query(createRequest("/"));
-          expect(response instanceof Response).toBe(true);
-          expect((response as Response).status).toBe(302);
-          expect((response as Response).headers.get("Location")).toBe(url);
-        }
-      });
-
-      it("should handle 404 navigations", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context = await query(createRequest("/not/found"));
-
-        expect(context).toMatchObject({
-          loaderData: {},
-          actionData: null,
-          errors: {
-            index: new ErrorResponseImpl(
-              404,
-              "Not Found",
-              new Error('No route matches URL "/not/found"'),
-              true
-            ),
-          },
-          matches: [{ route: { id: "index" } }],
-        });
-      });
-
-      it("should handle load error responses", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context;
-
-        // Error handled by child
-        context = await query(createRequest("/parent/error-boundary"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            parent: "PARENT LOADER",
-          },
-          errors: {
-            errorBoundary: "ERROR BOUNDARY LOADER ERROR",
-          },
-          matches: [
-            { route: { id: "parent" } },
-            { route: { id: "errorBoundary" } },
-          ],
-        });
-
-        // Error propagates to parent
-        context = await query(createRequest("/parent/error"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            parent: "PARENT LOADER",
-          },
-          errors: {
-            parent: "ERROR LOADER ERROR",
-          },
-          matches: [{ route: { id: "parent" } }, { route: { id: "error" } }],
-        });
-      });
-
-      it("should handle submit error responses", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let context;
-
-        // Error handled by child
-        context = await query(createSubmitRequest("/parent/error-boundary"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {
-            parent: "PARENT LOADER",
-          },
-          errors: {
-            errorBoundary: "ERROR BOUNDARY ACTION ERROR",
-          },
-          matches: [
-            { route: { id: "parent" } },
-            { route: { id: "errorBoundary" } },
-          ],
-        });
-
-        // Error propagates to parent
-        context = await query(createSubmitRequest("/parent/error"));
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {},
-          errors: {
-            parent: "ERROR ACTION ERROR",
-          },
-          matches: [{ route: { id: "parent" } }, { route: { id: "error" } }],
-        });
-      });
-
-      it("should handle multiple errors at separate boundaries", async () => {
-        let routes = [
-          {
-            id: "root",
-            path: "/",
-            loader: () => Promise.reject("ROOT"),
-            hasErrorBoundary: true,
-            children: [
-              {
-                id: "child",
-                path: "child",
-                loader: () => Promise.reject("CHILD"),
-                hasErrorBoundary: true,
-              },
-            ],
-          },
-        ];
-
-        let { query } = createStaticHandler(routes);
-        let context;
-
-        context = await query(createRequest("/child"));
-        expect(context.errors).toEqual({
-          root: "ROOT",
-          child: "CHILD",
-        });
-      });
-
-      it("should handle multiple errors at the same boundary", async () => {
-        let routes = [
-          {
-            id: "root",
-            path: "/",
-            loader: () => Promise.reject("ROOT"),
-            hasErrorBoundary: true,
-            children: [
-              {
-                id: "child",
-                path: "child",
-                loader: () => Promise.reject("CHILD"),
-              },
-            ],
-          },
-        ];
-
-        let { query } = createStaticHandler(routes);
-        let context;
-
-        context = await query(createRequest("/child"));
-        expect(context.errors).toEqual({
-          // higher error value wins
-          root: "ROOT",
-        });
-      });
-
-      it("should handle aborted load requests", async () => {
-        let dfd = createDeferred();
-        let controller = new AbortController();
-        let { query } = createStaticHandler([
-          {
-            id: "root",
-            path: "/path",
-            loader: () => dfd.promise,
-          },
-        ]);
-        let request = createRequest("/path?key=value", {
-          signal: controller.signal,
-        });
-        let e;
-        try {
-          let contextPromise = query(request);
-          controller.abort();
-          // This should resolve even though we never resolved the loader
-          await contextPromise;
-        } catch (_e) {
-          e = _e;
-        }
-        expect(e).toMatchInlineSnapshot(
-          `[Error: query() call aborted: GET http://localhost/path?key=value]`
-        );
-      });
-
-      it("should handle aborted submit requests", async () => {
-        let dfd = createDeferred();
-        let controller = new AbortController();
-        let { query } = createStaticHandler([
-          {
-            id: "root",
-            path: "/path",
-            action: () => dfd.promise,
-          },
-        ]);
-        let request = createSubmitRequest("/path?key=value", {
-          signal: controller.signal,
-        });
-        let e;
-        try {
-          let contextPromise = query(request);
-          controller.abort();
-          // This should resolve even though we never resolved the loader
-          await contextPromise;
-        } catch (_e) {
-          e = _e;
-        }
-        expect(e).toMatchInlineSnapshot(
-          `[Error: query() call aborted: POST http://localhost/path?key=value]`
-        );
-      });
-
-      it("should assign signals to requests by default (per the", async () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-        let request = createRequest("/", { signal: undefined });
-        let context = await query(request);
-        expect((context as StaticHandlerContext).loaderData.index).toBe(
-          "INDEX LOADER"
-        );
-      });
-
-      it("should handle not found action submissions with a 405 error", async () => {
-        let { query } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-          },
-        ]);
-        let request = createSubmitRequest("/");
-        let context = await query(request);
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {},
-          errors: {
-            root: new ErrorResponseImpl(
-              405,
-              "Method Not Allowed",
-              new Error(
-                'You made a POST request to "/" but did not provide an `action` ' +
-                  'for route "root", so there is no way to handle the request.'
-              ),
-              true
-            ),
-          },
-          matches: [{ route: { id: "root" } }],
-        });
-      });
-
-      it("should handle unsupported methods with a 405 error", async () => {
-        let { query } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-          },
-        ]);
-        let request = createRequest("/", { method: "OPTIONS" });
-        let context = await query(request);
-        expect(context).toMatchObject({
-          actionData: null,
-          loaderData: {},
-          errors: {
-            root: new ErrorResponseImpl(
-              405,
-              "Method Not Allowed",
-              new Error('Invalid request method "OPTIONS"'),
-              true
-            ),
-          },
-          matches: [{ route: { id: "root" } }],
-        });
-      });
-
-      it("should send proper arguments to loaders", async () => {
-        let rootLoaderStub = jest.fn(() => "ROOT");
-        let childLoaderStub = jest.fn(() => "CHILD");
-        let { query } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-            loader: rootLoaderStub,
-            children: [
-              {
-                id: "child",
-                path: "child",
-                loader: childLoaderStub,
-              },
-            ],
-          },
-        ]);
-        await query(createRequest("/child"));
-
-        // @ts-expect-error
-        let rootLoaderRequest = rootLoaderStub.mock.calls[0][0]?.request;
-        // @ts-expect-error
-        let childLoaderRequest = childLoaderStub.mock.calls[0][0]?.request;
-        expect(rootLoaderRequest.method).toBe("GET");
-        expect(rootLoaderRequest.url).toBe("http://localhost/child");
-        expect(childLoaderRequest.method).toBe("GET");
-        expect(childLoaderRequest.url).toBe("http://localhost/child");
-      });
-
-      it("should send proper arguments to actions", async () => {
-        let actionStub = jest.fn(() => "ACTION");
-        let rootLoaderStub = jest.fn(() => "ROOT");
-        let childLoaderStub = jest.fn(() => "CHILD");
-        let { query } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-            loader: rootLoaderStub,
-            children: [
-              {
-                id: "child",
-                path: "child",
-                action: actionStub,
-                loader: childLoaderStub,
-              },
-            ],
-          },
-        ]);
-        await query(
-          createSubmitRequest("/child", {
-            headers: {
-              test: "value",
-            },
-          })
-        );
-
-        // @ts-expect-error
-        let actionRequest = actionStub.mock.calls[0][0]?.request;
-        expect(actionRequest.method).toBe("POST");
-        expect(actionRequest.url).toBe("http://localhost/child");
-        expect(actionRequest.headers.get("Content-Type")).toBe(
-          "application/x-www-form-urlencoded;charset=UTF-8"
-        );
-        expect((await actionRequest.formData()).get("key")).toBe("value");
-
-        // @ts-expect-error
-        let rootLoaderRequest = rootLoaderStub.mock.calls[0][0]?.request;
-        // @ts-expect-error
-        let childLoaderRequest = childLoaderStub.mock.calls[0][0]?.request;
-        expect(rootLoaderRequest.method).toBe("GET");
-        expect(rootLoaderRequest.url).toBe("http://localhost/child");
-        expect(rootLoaderRequest.headers.get("test")).toBe("value");
-        expect(await rootLoaderRequest.text()).toBe("");
-        expect(childLoaderRequest.method).toBe("GET");
-        expect(childLoaderRequest.url).toBe("http://localhost/child");
-        expect(childLoaderRequest.headers.get("test")).toBe("value");
-        // Can't re-read body here since it's the same request as the root
-      });
-
-      it("should support a requestContext passed to loaders and actions", async () => {
-        let requestContext = { sessionId: "12345" };
-        let rootStub = jest.fn(() => "ROOT");
-        let childStub = jest.fn(() => "CHILD");
-        let actionStub = jest.fn(() => "CHILD ACTION");
-        let arg = (s) => s.mock.calls[0][0];
-        let { query } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-            loader: rootStub,
-            children: [
-              {
-                id: "child",
-                path: "child",
-                action: actionStub,
-                loader: childStub,
-              },
-            ],
-          },
-        ]);
-
-        await query(createRequest("/child"), { requestContext });
-        expect(arg(rootStub).context.sessionId).toBe("12345");
-        expect(arg(childStub).context.sessionId).toBe("12345");
-
-        actionStub.mockClear();
-        rootStub.mockClear();
-        childStub.mockClear();
-
-        await query(createSubmitRequest("/child"), { requestContext });
-        expect(arg(actionStub).context.sessionId).toBe("12345");
-        expect(arg(rootStub).context.sessionId).toBe("12345");
-        expect(arg(childStub).context.sessionId).toBe("12345");
-      });
-
-      describe("deferred", () => {
-        let { query } = createStaticHandler(SSR_ROUTES);
-
-        it("should return DeferredData on symbol", async () => {
-          let context = (await query(
-            createRequest("/parent/deferred")
-          )) as StaticHandlerContext;
-          expect(context).toMatchObject({
-            loaderData: {
-              parent: "PARENT LOADER",
-              deferred: {
-                critical: "loader",
-                lazy: expect.trackedPromise(),
-              },
-            },
-            activeDeferreds: {
-              deferred: expect.deferredData(false),
-            },
-          });
-          await new Promise((r) => setTimeout(r, 10));
-          expect(context).toMatchObject({
-            loaderData: {
-              parent: "PARENT LOADER",
-              deferred: {
-                critical: "loader",
-                lazy: expect.trackedPromise("lazy"),
-              },
-            },
-            activeDeferreds: {
-              deferred: expect.deferredData(true),
-            },
-          });
-        });
-
-        it("should return rejected DeferredData on symbol", async () => {
-          let context = (await query(
-            createRequest("/parent/deferred?reject")
-          )) as StaticHandlerContext;
-          expect(context).toMatchObject({
-            loaderData: {
-              parent: "PARENT LOADER",
-              deferred: {
-                critical: "loader",
-                lazy: expect.trackedPromise(),
-              },
-            },
-            activeDeferreds: {
-              deferred: expect.deferredData(false),
-            },
-          });
-          await new Promise((r) => setTimeout(r, 10));
-          expect(context).toMatchObject({
-            loaderData: {
-              parent: "PARENT LOADER",
-              deferred: {
-                critical: "loader",
-                lazy: expect.trackedPromise(undefined, new Error("broken!")),
-              },
-            },
-            activeDeferreds: {
-              deferred: expect.deferredData(true),
-            },
-          });
-        });
-
-        it("should return rejected DeferredData on symbol for resolved undefined", async () => {
-          let context = (await query(
-            createRequest("/parent/deferred?undefined")
-          )) as StaticHandlerContext;
-          expect(context).toMatchObject({
-            loaderData: {
-              parent: "PARENT LOADER",
-              deferred: {
-                critical: "loader",
-                lazy: expect.trackedPromise(),
-              },
-            },
-            activeDeferreds: {
-              deferred: expect.deferredData(false),
-            },
-          });
-          await new Promise((r) => setTimeout(r, 10));
-          expect(context).toMatchObject({
-            loaderData: {
-              parent: "PARENT LOADER",
-              deferred: {
-                critical: "loader",
-                lazy: expect.trackedPromise(
-                  null,
-                  new Error(
-                    `Deferred data for key "lazy" resolved/rejected with \`undefined\`, you must resolve/reject with a value or \`null\`.`
-                  )
-                ),
-              },
-            },
-            activeDeferreds: {
-              deferred: expect.deferredData(true),
-            },
-          });
-        });
-
-        it("should return DeferredData on symbol with status + headers", async () => {
-          let context = (await query(
-            createRequest("/parent/deferred?status")
-          )) as StaticHandlerContext;
-          expect(context).toMatchObject({
-            loaderData: {
-              parent: "PARENT LOADER",
-              deferred: {
-                critical: "loader",
-                lazy: expect.trackedPromise(),
-              },
-            },
-            activeDeferreds: {
-              deferred: expect.deferredData(false, 201, {
-                "x-custom": "yes",
-              }),
-            },
-          });
-          await new Promise((r) => setTimeout(r, 10));
-          expect(context).toMatchObject({
-            loaderData: {
-              parent: "PARENT LOADER",
-              deferred: {
-                critical: "loader",
-                lazy: expect.trackedPromise("lazy"),
-              },
-            },
-            activeDeferreds: {
-              deferred: expect.deferredData(true, 201, {
-                "x-custom": "yes",
-              }),
-            },
-            statusCode: 201,
-            loaderHeaders: {
-              deferred: new Headers({ "x-custom": "yes" }),
-            },
-          });
-        });
-
-        it("does not support deferred on submissions", async () => {
-          let context = (await query(
-            createSubmitRequest("/parent/deferred")
-          )) as StaticHandlerContext;
-          expect(context.actionData).toEqual(null);
-          expect(context.loaderData).toEqual({
-            parent: null,
-            deferred: null,
-          });
-          expect(context.activeDeferreds).toEqual(null);
-          expect(context.errors).toEqual({
-            parent: new ErrorResponseImpl(
-              400,
-              "Bad Request",
-              new Error("defer() is not supported in actions"),
-              true
-            ),
-          });
-        });
-      });
-
-      describe("statusCode", () => {
-        it("should expose a 200 status code by default", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-            },
-          ]);
-          let context = (await query(
-            createRequest("/")
-          )) as StaticHandlerContext;
-          expect(context.statusCode).toBe(200);
-        });
-
-        it("should expose a 500 status code on loader errors", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => json({ data: "ROOT" }, { status: 201 }),
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => {
-                    throw new Error("💥");
-                  },
-                },
-              ],
-            },
-          ]);
-          let context = (await query(
-            createRequest("/")
-          )) as StaticHandlerContext;
-          expect(context.statusCode).toBe(500);
-        });
-
-        it("should expose a 500 status code on action errors", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => json({ data: "ROOT" }, { status: 201 }),
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => json({ data: "CHILD" }, { status: 202 }),
-                  action: () => {
-                    throw new Error("💥");
-                  },
-                },
-              ],
-            },
-          ]);
-          let context = (await query(
-            createSubmitRequest("/?index")
-          )) as StaticHandlerContext;
-          expect(context.statusCode).toBe(500);
-        });
-
-        it("should expose a 4xx status code on thrown loader responses", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => json({ data: "ROOT" }, { status: 201 }),
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => {
-                    throw new Response(null, { status: 400 });
-                  },
-                },
-              ],
-            },
-          ]);
-          let context = (await query(
-            createRequest("/")
-          )) as StaticHandlerContext;
-          expect(context.statusCode).toBe(400);
-        });
-
-        it("should expose a 4xx status code on thrown action responses", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => json({ data: "ROOT" }, { status: 201 }),
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => json({ data: "CHILD" }, { status: 202 }),
-                  action: () => {
-                    throw new Response(null, { status: 400 });
-                  },
-                },
-              ],
-            },
-          ]);
-          let context = (await query(
-            createSubmitRequest("/?index")
-          )) as StaticHandlerContext;
-          expect(context.statusCode).toBe(400);
-        });
-
-        it("should expose the action status on submissions", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => json({ data: "ROOT" }, { status: 201 }),
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => json({ data: "ROOT" }, { status: 202 }),
-                  action: () => json({ data: "ROOT" }, { status: 203 }),
-                },
-              ],
-            },
-          ]);
-          let context = (await query(
-            createSubmitRequest("/?index")
-          )) as StaticHandlerContext;
-          expect(context.statusCode).toBe(203);
-        });
-
-        it("should expose the deepest 2xx status", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => json({ data: "ROOT" }, { status: 201 }),
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => json({ data: "ROOT" }, { status: 202 }),
-                },
-              ],
-            },
-          ]);
-          let context = (await query(
-            createRequest("/")
-          )) as StaticHandlerContext;
-          expect(context.statusCode).toBe(202);
-        });
-
-        it("should expose the shallowest 4xx/5xx status", async () => {
-          let context;
-          let query: StaticHandler["query"];
-
-          query = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => {
-                throw new Response(null, { status: 400 });
-              },
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => {
-                    throw new Response(null, { status: 401 });
-                  },
-                },
-              ],
-            },
-          ]).query;
-          context = (await query(createRequest("/"))) as StaticHandlerContext;
-          expect(context.statusCode).toBe(400);
-
-          query = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => {
-                throw new Response(null, { status: 400 });
-              },
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => {
-                    throw new Response(null, { status: 500 });
-                  },
-                },
-              ],
-            },
-          ]).query;
-          context = (await query(createRequest("/"))) as StaticHandlerContext;
-          expect(context.statusCode).toBe(400);
-
-          query = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => {
-                throw new Response(null, { status: 400 });
-              },
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => {
-                    throw new Error("💥");
-                  },
-                },
-              ],
-            },
-          ]).query;
-          context = (await query(createRequest("/"))) as StaticHandlerContext;
-          expect(context.statusCode).toBe(400);
-        });
-      });
-
-      describe("headers", () => {
-        it("should expose headers from action/loader responses", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => new Response(null, { headers: { two: "2" } }),
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  action: () => new Response(null, { headers: { one: "1" } }),
-                  loader: () => new Response(null, { headers: { three: "3" } }),
-                },
-              ],
-            },
-          ]);
-          let context = (await query(
-            createSubmitRequest("/?index")
-          )) as StaticHandlerContext;
-          expect(Array.from(context.actionHeaders.child.entries())).toEqual([
-            ["one", "1"],
-          ]);
-          expect(Array.from(context.loaderHeaders.root.entries())).toEqual([
-            ["two", "2"],
-          ]);
-          expect(Array.from(context.loaderHeaders.child.entries())).toEqual([
-            ["three", "3"],
-          ]);
-        });
-
-        it("should expose headers from loader error responses", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => new Response(null, { headers: { one: "1" } }),
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  loader: () => {
-                    throw new Response(null, { headers: { two: "2" } });
-                  },
-                },
-              ],
-            },
-          ]);
-          let context = (await query(
-            createRequest("/")
-          )) as StaticHandlerContext;
-          expect(Array.from(context.loaderHeaders.root.entries())).toEqual([
-            ["one", "1"],
-          ]);
-          expect(Array.from(context.loaderHeaders.child.entries())).toEqual([
-            ["two", "2"],
-          ]);
-        });
-
-        it("should expose headers from action error responses", async () => {
-          let { query } = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              children: [
-                {
-                  id: "child",
-                  index: true,
-                  action: () => {
-                    throw new Response(null, { headers: { one: "1" } });
-                  },
-                },
-              ],
-            },
-          ]);
-          let context = (await query(
-            createSubmitRequest("/?index")
-          )) as StaticHandlerContext;
-          expect(Array.from(context.actionHeaders.child.entries())).toEqual([
-            ["one", "1"],
-          ]);
-        });
-      });
-    });
-
-    describe("singular route requests", () => {
-      function setupFlexRouteTest() {
-        function queryRoute(
-          req: Request,
-          routeId: string,
-          type: "loader" | "action",
-          data: any,
-          isError = false
-        ) {
-          let handler = createStaticHandler([
-            {
-              id: "flex",
-              path: "/flex",
-              [type]: () =>
-                isError ? Promise.reject(data) : Promise.resolve(data),
-            },
-          ]);
-          return handler.queryRoute(req, { routeId });
-        }
-
-        return {
-          resolveLoader(data: any) {
-            return queryRoute(
-              createRequest("/flex"),
-              "flex",
-              "loader",
-              data,
-              false
-            );
-          },
-          rejectLoader(data: any) {
-            return queryRoute(
-              createRequest("/flex"),
-              "flex",
-              "loader",
-              data,
-              true
-            );
-          },
-          resolveAction(data: any) {
-            return queryRoute(
-              createSubmitRequest("/flex"),
-              "flex",
-              "action",
-              data,
-              false
-            );
-          },
-          rejectAction(data: any) {
-            return queryRoute(
-              createSubmitRequest("/flex"),
-              "flex",
-              "action",
-              data,
-              true
-            );
-          },
-        };
-      }
-
-      it("should match routes automatically if no routeId is provided", async () => {
-        let { queryRoute } = createStaticHandler(SSR_ROUTES);
-        let data;
-
-        data = await queryRoute(createRequest("/parent"));
-        expect(data).toBe("PARENT LOADER");
-
-        data = await queryRoute(createRequest("/parent?index"));
-        expect(data).toBe("PARENT INDEX LOADER");
-
-        data = await queryRoute(createRequest("/parent/child"), {
-          routeId: "child",
-        });
-        expect(data).toBe("CHILD LOADER");
-      });
-
-      it("should support HEAD requests", async () => {
-        let { queryRoute } = createStaticHandler(SSR_ROUTES);
-        let data = await queryRoute(
-          createRequest("/parent", { method: "HEAD" })
-        );
-        expect(data).toBe("PARENT LOADER");
-      });
-
-      it("should support OPTIONS requests", async () => {
-        let { queryRoute } = createStaticHandler(SSR_ROUTES);
-        let data = await queryRoute(
-          createRequest("/parent", { method: "OPTIONS" })
-        );
-        expect(data).toBe("PARENT LOADER");
-      });
-
-      it("should support singular route load navigations (primitives)", async () => {
-        let { queryRoute } = createStaticHandler(SSR_ROUTES);
-        let data;
-
-        // Layout route
-        data = await queryRoute(createRequest("/parent"), {
-          routeId: "parent",
-        });
-        expect(data).toBe("PARENT LOADER");
-
-        // Index route
-        data = await queryRoute(createRequest("/parent"), {
-          routeId: "parentIndex",
-        });
-        expect(data).toBe("PARENT INDEX LOADER");
-
-        // Parent in nested route
-        data = await queryRoute(createRequest("/parent/child"), {
-          routeId: "parent",
-        });
-        expect(data).toBe("PARENT LOADER");
-
-        // Child in nested route
-        data = await queryRoute(createRequest("/parent/child"), {
-          routeId: "child",
-        });
-        expect(data).toBe("CHILD LOADER");
-
-        // Non-undefined falsey values should count
-        let T = setupFlexRouteTest();
-        data = await T.resolveLoader(null);
-        expect(data).toBeNull();
-        data = await T.resolveLoader(false);
-        expect(data).toBe(false);
-        data = await T.resolveLoader("");
-        expect(data).toBe("");
-      });
-
-      it("should support singular route load navigations (Responses)", async () => {
-        /* eslint-disable jest/no-conditional-expect */
-        let T = setupFlexRouteTest();
-        let data;
-
-        // When Responses are returned or thrown, it should always resolve the
-        // raw Response from queryRoute
-
-        // Returned Success Response
-        data = await T.resolveLoader(new Response("Created!", { status: 201 }));
-        expect(data.status).toBe(201);
-        expect(await data.text()).toBe("Created!");
-
-        // Thrown Success Response
-        try {
-          await T.rejectLoader(new Response("Created!", { status: 201 }));
-          expect(false).toBe(true);
-        } catch (data) {
-          expect(data.status).toBe(201);
-          expect(await data.text()).toBe("Created!");
-        }
-
-        // Returned Redirect Response
-        data = await T.resolveLoader(
-          new Response(null, {
-            status: 302,
-            headers: { Location: "/" },
-          })
-        );
-        expect(data.status).toBe(302);
-        expect(data.headers.get("Location")).toBe("/");
-
-        // Thrown Redirect Response
-        data = await T.rejectLoader(
-          new Response(null, {
-            status: 301,
-            headers: { Location: "/" },
-          })
-        );
-        expect(data.status).toBe(301);
-        expect(data.headers.get("Location")).toBe("/");
-
-        // Returned Error Response
-        data = await T.resolveLoader(new Response("Why?", { status: 400 }));
-        expect(data.status).toBe(400);
-        expect(await data.text()).toBe("Why?");
-
-        // Thrown Error Response
-        try {
-          await T.rejectLoader(new Response("Oh no!", { status: 401 }));
-          expect(false).toBe(true);
-        } catch (data) {
-          expect(data.status).toBe(401);
-          expect(await data.text()).toBe("Oh no!");
-        }
-        /* eslint-enable jest/no-conditional-expect */
-      });
-
-      it("should support singular route load navigations (Errors)", async () => {
-        let T = setupFlexRouteTest();
-        let data;
-
-        // Returned Error instance is treated as data since it was not thrown
-        data = await T.resolveLoader(new Error("Why?"));
-        expect(data).toEqual(new Error("Why?"));
-
-        // Anything thrown (Error instance or not) will throw from queryRoute
-        // so we know to handle it as an errorPath in the server.  Generally
-        // though in queryRoute, we would expect responses to be coming back -
-        // not
-
-        // Thrown Error
-        try {
-          await T.rejectLoader(new Error("Oh no!"));
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toEqual(new Error("Oh no!"));
-
-        // Thrown non-Error
-        try {
-          await T.rejectLoader("This is weird?");
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toEqual("This is weird?");
-
-        // Non-undefined falsey values should count
-        try {
-          await T.rejectLoader(null);
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toBeNull();
-        try {
-          await T.rejectLoader(false);
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toBe(false);
-        try {
-          await T.rejectLoader("");
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toBe("");
-      });
-
-      it("should support singular route load navigations (with a basename)", async () => {
-        let { queryRoute } = createStaticHandler(SSR_ROUTES, {
-          basename: "/base",
-        });
-        let data;
-
-        // Layout route
-        data = await queryRoute(createRequest("/base/parent"), {
-          routeId: "parent",
-        });
-        expect(data).toBe("PARENT LOADER");
-
-        // Index route
-        data = await queryRoute(createRequest("/base/parent"), {
-          routeId: "parentIndex",
-        });
-        expect(data).toBe("PARENT INDEX LOADER");
-
-        // Parent in nested route
-        data = await queryRoute(createRequest("/base/parent/child"), {
-          routeId: "parent",
-        });
-        expect(data).toBe("PARENT LOADER");
-
-        // Child in nested route
-        data = await queryRoute(createRequest("/base/parent/child"), {
-          routeId: "child",
-        });
-        expect(data).toBe("CHILD LOADER");
-
-        // Non-undefined falsey values should count
-        let T = setupFlexRouteTest();
-        data = await T.resolveLoader(null);
-        expect(data).toBeNull();
-        data = await T.resolveLoader(false);
-        expect(data).toBe(false);
-        data = await T.resolveLoader("");
-        expect(data).toBe("");
-      });
-
-      it("should support singular route submit navigations (primitives)", async () => {
-        let { queryRoute } = createStaticHandler(SSR_ROUTES);
-        let data;
-
-        // Layout route
-        data = await queryRoute(createSubmitRequest("/parent"), {
-          routeId: "parent",
-        });
-        expect(data).toBe("PARENT ACTION");
-
-        // Index route
-        data = await queryRoute(createSubmitRequest("/parent"), {
-          routeId: "parentIndex",
-        });
-        expect(data).toBe("PARENT INDEX ACTION");
-
-        // Parent in nested route
-        data = await queryRoute(createSubmitRequest("/parent/child"), {
-          routeId: "parent",
-        });
-        expect(data).toBe("PARENT ACTION");
-
-        // Child in nested route
-        data = await queryRoute(createSubmitRequest("/parent/child"), {
-          routeId: "child",
-        });
-        expect(data).toBe("CHILD ACTION");
-
-        // Non-undefined falsey values should count
-        let T = setupFlexRouteTest();
-        data = await T.resolveAction(null);
-        expect(data).toBeNull();
-        data = await T.resolveAction(false);
-        expect(data).toBe(false);
-        data = await T.resolveAction("");
-        expect(data).toBe("");
-      });
-
-      it("should support alternative submission methods", async () => {
-        let { queryRoute } = createStaticHandler(SSR_ROUTES);
-        let data;
-
-        data = await queryRoute(
-          createSubmitRequest("/parent", { method: "PUT" }),
-          { routeId: "parent" }
-        );
-        expect(data).toBe("PARENT ACTION");
-
-        data = await queryRoute(
-          createSubmitRequest("/parent", { method: "PATCH" }),
-          { routeId: "parent" }
-        );
-        expect(data).toBe("PARENT ACTION");
-
-        data = await queryRoute(
-          createSubmitRequest("/parent", { method: "DELETE" }),
-          { routeId: "parent" }
-        );
-        expect(data).toBe("PARENT ACTION");
-      });
-
-      it("should support singular route submit navigations (Responses)", async () => {
-        /* eslint-disable jest/no-conditional-expect */
-        let T = setupFlexRouteTest();
-        let data;
-
-        // When Responses are returned or thrown, it should always resolve the
-        // raw Response from queryRoute
-
-        // Returned Success Response
-        data = await T.resolveAction(new Response("Created!", { status: 201 }));
-        expect(data.status).toBe(201);
-        expect(await data.text()).toBe("Created!");
-
-        // Thrown Success Response
-        try {
-          await T.rejectAction(new Response("Created!", { status: 201 }));
-          expect(false).toBe(true);
-        } catch (data) {
-          expect(data.status).toBe(201);
-          expect(await data.text()).toBe("Created!");
-        }
-
-        // Returned Redirect Response
-        data = await T.resolveAction(
-          new Response(null, {
-            status: 302,
-            headers: { Location: "/" },
-          })
-        );
-        expect(data.status).toBe(302);
-        expect(data.headers.get("Location")).toBe("/");
-
-        // Thrown Redirect Response
-        data = await T.rejectAction(
-          new Response(null, {
-            status: 301,
-            headers: { Location: "/" },
-          })
-        );
-        expect(data.status).toBe(301);
-        expect(data.headers.get("Location")).toBe("/");
-
-        // Returned Error Response
-        data = await T.resolveAction(new Response("Why?", { status: 400 }));
-        expect(data.status).toBe(400);
-        expect(await data.text()).toBe("Why?");
-
-        // Thrown Error Response
-        try {
-          await T.rejectAction(new Response("Oh no!", { status: 401 }));
-          expect(false).toBe(true);
-        } catch (data) {
-          expect(data.status).toBe(401);
-          expect(await data.text()).toBe("Oh no!");
-        }
-        /* eslint-enable jest/no-conditional-expect */
-      });
-
-      it("should support singular route submit navigations (Errors)", async () => {
-        let T = setupFlexRouteTest();
-        let data;
-
-        // Returned Error instance is treated as data since it was not thrown
-        data = await T.resolveAction(new Error("Why?"));
-        expect(data).toEqual(new Error("Why?"));
-
-        // Anything thrown (Error instance or not) will throw from queryRoute
-        // so we know to handle it as an errorPath in the server.  Generally
-        // though in queryRoute, we would expect responses to be coming back -
-        // not
-
-        // Thrown Error
-        try {
-          await T.rejectAction(new Error("Oh no!"));
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toEqual(new Error("Oh no!"));
-
-        // Thrown non-Error
-        try {
-          await T.rejectAction("This is weird?");
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toEqual("This is weird?");
-
-        // Non-undefined falsey values should count
-        try {
-          await T.rejectAction(null);
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toBeNull();
-        try {
-          await T.rejectAction(false);
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toBe(false);
-        try {
-          await T.rejectAction("");
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toBe("");
-      });
-
-      it("should error if an action/loader returns undefined", async () => {
-        let T = setupFlexRouteTest();
-        let data;
-
-        try {
-          data = await T.resolveLoader(undefined);
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toEqual(
-          new Error(
-            'You defined a loader for route "flex" but didn\'t return anything ' +
-              "from your `loader` function. Please return a value or `null`."
-          )
-        );
-
-        try {
-          data = await T.resolveAction(undefined);
-        } catch (e) {
-          data = e;
-        }
-        expect(data).toEqual(
-          new Error(
-            'You defined an action for route "flex" but didn\'t return anything ' +
-              "from your `action` function. Please return a value or `null`."
-          )
-        );
-      });
-
-      it("should handle relative redirect responses (loader)", async () => {
-        let { queryRoute } = createStaticHandler([
-          {
-            path: "/",
-            children: [
-              {
-                path: "parent",
-                children: [
-                  {
-                    id: "child",
-                    path: "child",
-                    loader: () => redirect(".."),
-                  },
-                ],
-              },
-            ],
-          },
-        ]);
-        let response = await queryRoute(createRequest("/parent/child"), {
-          routeId: "child",
-        });
-        expect(response instanceof Response).toBe(true);
-        expect((response as Response).status).toBe(302);
-        expect((response as Response).headers.get("Location")).toBe("/parent");
-      });
-
-      it("should handle relative redirect responses (action)", async () => {
-        let { queryRoute } = createStaticHandler([
-          {
-            path: "/",
-            children: [
-              {
-                path: "parent",
-                children: [
-                  {
-                    id: "child",
-                    path: "child",
-                    action: () => redirect(".."),
-                  },
-                ],
-              },
-            ],
-          },
-        ]);
-        let response = await queryRoute(createSubmitRequest("/parent/child"), {
-          routeId: "child",
-        });
-        expect(response instanceof Response).toBe(true);
-        expect((response as Response).status).toBe(302);
-        expect((response as Response).headers.get("Location")).toBe("/parent");
-      });
-
-      it("should handle absolute redirect Responses", async () => {
-        for (let url of ABSOLUTE_URLS) {
-          let handler = createStaticHandler([
-            {
-              id: "root",
-              path: "/",
-              loader: () => redirect(url),
-            },
-          ]);
-          let response = await handler.queryRoute(createRequest("/"), {
-            routeId: "root",
-          });
-          expect(response instanceof Response).toBe(true);
-          expect((response as Response).status).toBe(302);
-          expect((response as Response).headers.get("Location")).toBe(url);
-        }
-      });
-
-      it("should not unwrap responses returned from loaders", async () => {
-        let response = json({ key: "value" });
-        let { queryRoute } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-            loader: () => Promise.resolve(response),
-          },
-        ]);
-        let request = createRequest("/");
-        let data = await queryRoute(request, { routeId: "root" });
-        expect(data instanceof Response).toBe(true);
-        expect(await data.json()).toEqual({ key: "value" });
-      });
-
-      it("should not unwrap responses returned from actions", async () => {
-        let response = json({ key: "value" });
-        let { queryRoute } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-            action: () => Promise.resolve(response),
-          },
-        ]);
-        let request = createSubmitRequest("/");
-        let data = await queryRoute(request, { routeId: "root" });
-        expect(data instanceof Response).toBe(true);
-        expect(await data.json()).toEqual({ key: "value" });
-      });
-
-      it("should not unwrap responses thrown from loaders", async () => {
-        let response = json({ key: "value" });
-        let { queryRoute } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-            loader: () => Promise.reject(response),
-          },
-        ]);
-        let request = createRequest("/");
-        let data;
-        try {
-          await queryRoute(request, { routeId: "root" });
-        } catch (e) {
-          data = e;
-        }
-        expect(data instanceof Response).toBe(true);
-        expect(await data.json()).toEqual({ key: "value" });
-      });
-
-      it("should not unwrap responses thrown from actions", async () => {
-        let response = json({ key: "value" });
-        let { queryRoute } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-            action: () => Promise.reject(response),
-          },
-        ]);
-        let request = createSubmitRequest("/");
-        let data;
-        try {
-          await queryRoute(request, { routeId: "root" });
-        } catch (e) {
-          data = e;
-        }
-        expect(data instanceof Response).toBe(true);
-        expect(await data.json()).toEqual({ key: "value" });
-      });
-
-      it("should handle aborted load requests", async () => {
-        let dfd = createDeferred();
-        let controller = new AbortController();
-        let { queryRoute } = createStaticHandler([
-          {
-            id: "root",
-            path: "/path",
-            loader: () => dfd.promise,
-          },
-        ]);
-        let request = createRequest("/path?key=value", {
-          signal: controller.signal,
-        });
-        let e;
-        try {
-          let statePromise = queryRoute(request, { routeId: "root" });
-          controller.abort();
-          // This should resolve even though we never resolved the loader
-          await statePromise;
-        } catch (_e) {
-          e = _e;
-        }
-        expect(e).toMatchInlineSnapshot(
-          `[Error: queryRoute() call aborted: GET http://localhost/path?key=value]`
-        );
-      });
-
-      it("should handle aborted submit requests", async () => {
-        let dfd = createDeferred();
-        let controller = new AbortController();
-        let { queryRoute } = createStaticHandler([
-          {
-            id: "root",
-            path: "/path",
-            action: () => dfd.promise,
-          },
-        ]);
-        let request = createSubmitRequest("/path?key=value", {
-          signal: controller.signal,
-        });
-        let e;
-        try {
-          let statePromise = queryRoute(request, { routeId: "root" });
-          controller.abort();
-          // This should resolve even though we never resolved the loader
-          await statePromise;
-        } catch (_e) {
-          e = _e;
-        }
-        expect(e).toMatchInlineSnapshot(
-          `[Error: queryRoute() call aborted: POST http://localhost/path?key=value]`
-        );
-      });
-
-      it("should assign signals to requests by default (per the spec)", async () => {
-        let { queryRoute } = createStaticHandler(SSR_ROUTES);
-        let request = createRequest("/", { signal: undefined });
-        let data = await queryRoute(request, { routeId: "index" });
-        expect(data).toBe("INDEX LOADER");
-      });
-
-      it("should support a requestContext passed to loaders and actions", async () => {
-        let requestContext = { sessionId: "12345" };
-        let childStub = jest.fn(() => "CHILD");
-        let actionStub = jest.fn(() => "CHILD ACTION");
-        let arg = (s) => s.mock.calls[0][0];
-        let { queryRoute } = createStaticHandler([
-          {
-            path: "/",
-            children: [
-              {
-                id: "child",
-                path: "child",
-                action: actionStub,
-                loader: childStub,
-              },
-            ],
-          },
-        ]);
-
-        await queryRoute(createRequest("/child"), {
-          routeId: "child",
-          requestContext,
-        });
-        expect(arg(childStub).context.sessionId).toBe("12345");
-
-        await queryRoute(createSubmitRequest("/child"), {
-          routeId: "child",
-          requestContext,
-        });
-        expect(arg(actionStub).context.sessionId).toBe("12345");
-      });
-
-      describe("deferred", () => {
-        let { queryRoute } = createStaticHandler(SSR_ROUTES);
-
-        it("should return DeferredData on symbol", async () => {
-          let result = await queryRoute(createRequest("/parent/deferred"));
-          expect(result).toMatchObject({
-            critical: "loader",
-            lazy: expect.trackedPromise(),
-          });
-          expect(result[UNSAFE_DEFERRED_SYMBOL]).deferredData(false);
-          await new Promise((r) => setTimeout(r, 10));
-          expect(result).toMatchObject({
-            critical: "loader",
-            lazy: expect.trackedPromise("lazy"),
-          });
-          expect(result[UNSAFE_DEFERRED_SYMBOL]).deferredData(true);
-        });
-
-        it("should return rejected DeferredData on symbol", async () => {
-          let result = await queryRoute(
-            createRequest("/parent/deferred?reject")
-          );
-          expect(result).toMatchObject({
-            critical: "loader",
-            lazy: expect.trackedPromise(),
-          });
-          expect(result[UNSAFE_DEFERRED_SYMBOL]).deferredData(false);
-          await new Promise((r) => setTimeout(r, 10));
-          expect(result).toMatchObject({
-            critical: "loader",
-            lazy: expect.trackedPromise(null, new Error("broken!")),
-          });
-          expect(result[UNSAFE_DEFERRED_SYMBOL]).deferredData(true);
-        });
-
-        it("should return rejected DeferredData on symbol for resolved undefined", async () => {
-          let result = await queryRoute(
-            createRequest("/parent/deferred?undefined")
-          );
-          expect(result).toMatchObject({
-            critical: "loader",
-            lazy: expect.trackedPromise(),
-          });
-          expect(result[UNSAFE_DEFERRED_SYMBOL]).deferredData(false);
-          await new Promise((r) => setTimeout(r, 10));
-          expect(result).toMatchObject({
-            critical: "loader",
-            lazy: expect.trackedPromise(
-              null,
-              new Error(
-                `Deferred data for key "lazy" resolved/rejected with \`undefined\`, you must resolve/reject with a value or \`null\`.`
-              )
-            ),
-          });
-          expect(result[UNSAFE_DEFERRED_SYMBOL]).deferredData(true);
-        });
-
-        it("should return DeferredData on symbol with status + headers", async () => {
-          let result = await queryRoute(
-            createRequest("/parent/deferred?status")
-          );
-          expect(result).toMatchObject({
-            critical: "loader",
-            lazy: expect.trackedPromise(),
-          });
-          expect(result[UNSAFE_DEFERRED_SYMBOL]).deferredData(false, 201, {
-            "x-custom": "yes",
-          });
-          await new Promise((r) => setTimeout(r, 10));
-          expect(result).toMatchObject({
-            critical: "loader",
-            lazy: expect.trackedPromise("lazy"),
-          });
-          expect(result[UNSAFE_DEFERRED_SYMBOL]).deferredData(true, 201, {
-            "x-custom": "yes",
-          });
-        });
-
-        it("does not support deferred on submissions", async () => {
-          try {
-            await queryRoute(createSubmitRequest("/parent/deferred"));
-            expect(false).toBe(true);
-          } catch (e) {
-            // eslint-disable-next-line jest/no-conditional-expect
-            expect(e).toEqual(
-              new ErrorResponseImpl(
-                400,
-                "Bad Request",
-                new Error("defer() is not supported in actions"),
-                true
-              )
-            );
-          }
-        });
-      });
-
-      describe("Errors with Status Codes", () => {
-        /* eslint-disable jest/no-conditional-expect */
-        let { queryRoute } = createStaticHandler([
-          {
-            id: "root",
-            path: "/",
-          },
-        ]);
-
-        it("should handle not found paths with a 404 Response", async () => {
-          try {
-            await queryRoute(createRequest("/junk"));
-            expect(false).toBe(true);
-          } catch (data) {
-            expect(isRouteErrorResponse(data)).toBe(true);
-            expect(data.status).toBe(404);
-            expect(data.error).toEqual(
-              new Error('No route matches URL "/junk"')
-            );
-            expect(data.internal).toBe(true);
-          }
-
-          try {
-            await queryRoute(createSubmitRequest("/junk"));
-            expect(false).toBe(true);
-          } catch (data) {
-            expect(isRouteErrorResponse(data)).toBe(true);
-            expect(data.status).toBe(404);
-            expect(data.error).toEqual(
-              new Error('No route matches URL "/junk"')
-            );
-            expect(data.internal).toBe(true);
-          }
-        });
-
-        it("should handle not found routeIds with a 403 Response", async () => {
-          try {
-            await queryRoute(createRequest("/"), { routeId: "junk" });
-            expect(false).toBe(true);
-          } catch (data) {
-            expect(isRouteErrorResponse(data)).toBe(true);
-            expect(data.status).toBe(403);
-            expect(data.error).toEqual(
-              new Error('Route "junk" does not match URL "/"')
-            );
-            expect(data.internal).toBe(true);
-          }
-
-          try {
-            await queryRoute(createSubmitRequest("/"), { routeId: "junk" });
-            expect(false).toBe(true);
-          } catch (data) {
-            expect(isRouteErrorResponse(data)).toBe(true);
-            expect(data.status).toBe(403);
-            expect(data.error).toEqual(
-              new Error('Route "junk" does not match URL "/"')
-            );
-            expect(data.internal).toBe(true);
-          }
-        });
-
-        it("should handle missing loaders with a 400 Response", async () => {
-          try {
-            await queryRoute(createRequest("/"), { routeId: "root" });
-            expect(false).toBe(true);
-          } catch (data) {
-            expect(isRouteErrorResponse(data)).toBe(true);
-            expect(data.status).toBe(400);
-            expect(data.error).toEqual(
-              new Error(
-                'You made a GET request to "/" but did not provide a `loader` ' +
-                  'for route "root", so there is no way to handle the request.'
-              )
-            );
-            expect(data.internal).toBe(true);
-          }
-        });
-
-        it("should handle missing actions with a 405 Response", async () => {
-          try {
-            await queryRoute(createSubmitRequest("/"), { routeId: "root" });
-            expect(false).toBe(true);
-          } catch (data) {
-            expect(isRouteErrorResponse(data)).toBe(true);
-            expect(data.status).toBe(405);
-            expect(data.error).toEqual(
-              new Error(
-                'You made a POST request to "/" but did not provide an `action` ' +
-                  'for route "root", so there is no way to handle the request.'
-              )
-            );
-            expect(data.internal).toBe(true);
-          }
-        });
-
-        it("should handle unsupported methods with a 405 Response", async () => {
-          try {
-            await queryRoute(createRequest("/", { method: "CHICKEN" }), {
-              routeId: "root",
-            });
-            expect(false).toBe(true);
-          } catch (data) {
-            expect(isRouteErrorResponse(data)).toBe(true);
-            expect(data.status).toBe(405);
-            expect(data.error).toEqual(
-              new Error('Invalid request method "CHICKEN"')
-            );
-            expect(data.internal).toBe(true);
-          }
-        });
-
-        /* eslint-enable jest/no-conditional-expect */
-      });
-    });
-  });
-
   describe("routes updates", () => {
     it("should retain existing routes until revalidation completes on loader removal", async () => {
       let t = initializeTmTest();
@@ -16447,586 +12579,6 @@ describe("a router", () => {
       });
 
       cleanup(router);
-    });
-  });
-
-  describe("path resolution", () => {
-    describe("routing to self", () => {
-      // Utility that accepts children of /foo routes and executes the same
-      // routing tests starting at /foo/bar/?a=b#hash
-      function assertRoutingToSelf(fooChildren, expectedPath, expectIndex) {
-        const getRouter = () =>
-          createRouter({
-            routes: [
-              {
-                path: "/",
-                children: [
-                  {
-                    path: "foo",
-                    children: fooChildren,
-                  },
-                ],
-              },
-            ],
-            history: createMemoryHistory({
-              initialEntries: ["/foo/bar?a=1#hash"],
-            }),
-          }).initialize();
-
-        // Null should preserve the search/hash
-        let router = getRouter();
-        router.navigate(null, { fromRouteId: "activeRoute" });
-        expect(createPath(router.state.location)).toBe(
-          expectedPath + (expectIndex ? "?index&a=1#hash" : "?a=1#hash")
-        );
-        router.dispose();
-
-        // "." and "" should not preserve the search and hash
-        router = getRouter();
-        router.navigate(".", { fromRouteId: "activeRoute" });
-        expect(createPath(router.state.location)).toBe(
-          expectedPath + (expectIndex ? "?index" : "")
-        );
-        router.dispose();
-
-        router = getRouter();
-        router.navigate("", { fromRouteId: "activeRoute" });
-        expect(createPath(router.state.location)).toBe(
-          expectedPath + (expectIndex ? "?index" : "")
-        );
-        router.dispose();
-      }
-
-      /* eslint-disable jest/expect-expect */
-      it("from a static route", () => {
-        assertRoutingToSelf(
-          [
-            {
-              id: "activeRoute",
-              path: "bar",
-            },
-          ],
-          "/foo/bar",
-          false
-        );
-      });
-
-      it("from a layout route", () => {
-        assertRoutingToSelf(
-          [
-            {
-              id: "activeRoute",
-              path: "bar",
-              children: [
-                {
-                  index: true,
-                },
-              ],
-            },
-          ],
-          "/foo/bar",
-          false
-        );
-      });
-
-      it("from an index route", () => {
-        assertRoutingToSelf(
-          [
-            {
-              path: "bar",
-              children: [
-                {
-                  id: "activeRoute",
-                  index: true,
-                },
-              ],
-            },
-          ],
-          "/foo/bar",
-          true
-        );
-      });
-
-      it("from an index route with a path", () => {
-        assertRoutingToSelf(
-          [
-            {
-              id: "activeRoute",
-              path: "bar",
-              index: true,
-            },
-          ],
-          "/foo/bar",
-          true
-        );
-      });
-
-      it("from a dynamic param route", () => {
-        assertRoutingToSelf(
-          [
-            {
-              id: "activeRoute",
-              path: ":param",
-            },
-          ],
-          "/foo/bar",
-          false
-        );
-      });
-
-      it("from a splat route", () => {
-        assertRoutingToSelf(
-          [
-            {
-              id: "activeRoute",
-              path: "*",
-            },
-          ],
-          "/foo",
-          false
-        );
-      });
-      /* eslint-enable jest/expect-expect */
-    });
-
-    describe("routing to parent", () => {
-      function assertRoutingToParent(fooChildren) {
-        let router = createRouter({
-          routes: [
-            {
-              path: "/",
-              children: [
-                {
-                  path: "foo",
-                  children: fooChildren,
-                },
-              ],
-            },
-          ],
-          history: createMemoryHistory({
-            initialEntries: ["/foo/bar?a=1#hash"],
-          }),
-        }).initialize();
-
-        // Null should preserve the search/hash
-        router.navigate("..", { fromRouteId: "activeRoute" });
-        expect(createPath(router.state.location)).toBe("/foo");
-      }
-
-      /* eslint-disable jest/expect-expect */
-      it("from a static route", () => {
-        assertRoutingToParent([
-          {
-            id: "activeRoute",
-            path: "bar",
-          },
-        ]);
-      });
-
-      it("from a layout route", () => {
-        assertRoutingToParent([
-          {
-            id: "activeRoute",
-            path: "bar",
-            children: [
-              {
-                index: true,
-              },
-            ],
-          },
-        ]);
-      });
-
-      it("from an index route", () => {
-        assertRoutingToParent([
-          {
-            path: "bar",
-            children: [
-              {
-                id: "activeRoute",
-                index: true,
-              },
-            ],
-          },
-        ]);
-      });
-
-      it("from an index route with a path", () => {
-        assertRoutingToParent([
-          {
-            id: "activeRoute",
-            path: "bar",
-            index: true,
-          },
-        ]);
-      });
-
-      it("from a dynamic param route", () => {
-        assertRoutingToParent([
-          {
-            id: "activeRoute",
-            path: ":param",
-          },
-        ]);
-      });
-
-      it("from a splat route", () => {
-        assertRoutingToParent([
-          {
-            id: "activeRoute",
-            path: "*",
-          },
-        ]);
-      });
-      /* eslint-enable jest/expect-expect */
-    });
-
-    describe("routing to sibling", () => {
-      function assertRoutingToSibling(fooChildren) {
-        let router = createRouter({
-          routes: [
-            {
-              path: "/",
-              children: [
-                {
-                  path: "foo",
-                  children: [
-                    ...fooChildren,
-                    {
-                      path: "bar-sibling",
-                    },
-                  ],
-                },
-              ],
-            },
-          ],
-          history: createMemoryHistory({
-            initialEntries: ["/foo/bar?a=1#hash"],
-          }),
-        }).initialize();
-
-        // Null should preserve the search/hash
-        router.navigate("../bar-sibling", { fromRouteId: "activeRoute" });
-        expect(createPath(router.state.location)).toBe("/foo/bar-sibling");
-      }
-
-      /* eslint-disable jest/expect-expect */
-      it("from a static route", () => {
-        assertRoutingToSibling([
-          {
-            id: "activeRoute",
-            path: "bar",
-          },
-        ]);
-      });
-
-      it("from a layout route", () => {
-        assertRoutingToSibling([
-          {
-            id: "activeRoute",
-            path: "bar",
-            children: [
-              {
-                index: true,
-              },
-            ],
-          },
-        ]);
-      });
-
-      it("from an index route", () => {
-        assertRoutingToSibling([
-          {
-            path: "bar",
-            children: [
-              {
-                id: "activeRoute",
-                index: true,
-              },
-            ],
-          },
-        ]);
-      });
-
-      it("from an index route with a path", () => {
-        assertRoutingToSibling([
-          {
-            id: "activeRoute",
-            path: "bar",
-            index: true,
-          },
-        ]);
-      });
-
-      it("from a dynamic param route", () => {
-        assertRoutingToSibling([
-          {
-            id: "activeRoute",
-            path: ":param",
-          },
-        ]);
-      });
-
-      it("from a splat route", () => {
-        assertRoutingToSibling([
-          {
-            id: "activeRoute",
-            path: "*",
-          },
-        ]);
-      });
-      /* eslint-enable jest/expect-expect */
-    });
-
-    describe("routing to child", () => {
-      function assertRoutingToChild(fooChildren) {
-        const getRouter = () =>
-          createRouter({
-            routes: [
-              {
-                path: "/",
-                children: [
-                  {
-                    path: "foo",
-                    children: [...fooChildren],
-                  },
-                ],
-              },
-            ],
-            history: createMemoryHistory({
-              initialEntries: ["/foo/bar?a=1#hash"],
-            }),
-          }).initialize();
-
-        let router = getRouter();
-        router.navigate("baz", { fromRouteId: "activeRoute" });
-        expect(createPath(router.state.location)).toBe("/foo/bar/baz");
-        router.dispose();
-
-        router = getRouter();
-        router.navigate("./baz", { fromRouteId: "activeRoute" });
-        expect(createPath(router.state.location)).toBe("/foo/bar/baz");
-        router.dispose();
-      }
-
-      /* eslint-disable jest/expect-expect */
-      it("from a static route", () => {
-        assertRoutingToChild([
-          {
-            id: "activeRoute",
-            path: "bar",
-            children: [{ path: "baz" }],
-          },
-        ]);
-      });
-
-      it("from a layout route", () => {
-        assertRoutingToChild([
-          {
-            id: "activeRoute",
-            path: "bar",
-            children: [
-              {
-                index: true,
-              },
-              { path: "baz" },
-            ],
-          },
-        ]);
-      });
-
-      it("from a dynamic param route", () => {
-        assertRoutingToChild([
-          {
-            id: "activeRoute",
-            path: ":param",
-            children: [{ path: "baz" }],
-          },
-        ]);
-      });
-      /* eslint-enable jest/expect-expect */
-    });
-
-    it("resolves relative routes when using relative:path", () => {
-      let history = createMemoryHistory({
-        initialEntries: ["/a/b/c/d/e/f"],
-      });
-      let routes = [
-        {
-          id: "a",
-          path: "/a",
-          children: [
-            {
-              id: "bc",
-              path: "b/c",
-              children: [
-                {
-                  id: "de",
-                  path: "d/e",
-                  children: [
-                    {
-                      id: "f",
-                      path: "f",
-                    },
-                  ],
-                },
-              ],
-            },
-          ],
-        },
-      ];
-
-      // Navigating without relative:path
-      let router = createRouter({ routes, history }).initialize();
-      router.navigate("..");
-      expect(router.state.location.pathname).toBe("/a/b/c/d/e");
-      router.navigate("/a/b/c/d/e/f");
-
-      router.navigate("../..");
-      expect(router.state.location.pathname).toBe("/a/b/c");
-      router.navigate("/a/b/c/d/e/f");
-
-      // Navigating with relative:path
-      router.navigate("..", { relative: "path" });
-      expect(router.state.location.pathname).toBe("/a/b/c/d/e");
-      router.navigate("/a/b/c/d/e/f");
-
-      router.navigate("../..", { relative: "path" });
-      expect(router.state.location.pathname).toBe("/a/b/c/d");
-      router.navigate("/a/b/c/d/e/f");
-
-      router.dispose();
-    });
-
-    it("should not append ?index to get submission navigations to self from index route", () => {
-      let router = createRouter({
-        routes: [
-          {
-            path: "/",
-            children: [
-              {
-                path: "path",
-                children: [
-                  {
-                    id: "activeRouteId",
-                    index: true,
-                  },
-                ],
-              },
-            ],
-          },
-        ],
-        history: createMemoryHistory({ initialEntries: ["/path"] }),
-      }).initialize();
-
-      router.navigate(null, {
-        fromRouteId: "activeRouteId",
-        formData: createFormData({}),
-      });
-      expect(createPath(router.state.location)).toBe("/path");
-      expect(router.state.matches[2].route.index).toBe(true);
-
-      router.navigate(".", {
-        fromRouteId: "activeRouteId",
-        formData: createFormData({}),
-      });
-      expect(createPath(router.state.location)).toBe("/path");
-      expect(router.state.matches[2].route.index).toBe(true);
-
-      router.navigate("", {
-        fromRouteId: "activeRouteId",
-        formData: createFormData({}),
-      });
-      expect(createPath(router.state.location)).toBe("/path");
-      expect(router.state.matches[2].route.index).toBe(true);
-    });
-
-    it("handles pathless relative routing when a basename is present", () => {
-      let router = createRouter({
-        routes: [{ path: "/path" }],
-        future: { v7_prependBasename: true },
-        history: createMemoryHistory({ initialEntries: ["/base/path"] }),
-        basename: "/base",
-      }).initialize();
-
-      expect(createPath(router.state.location)).toBe("/base/path");
-      expect(router.state.matches[0].route.path).toBe("/path");
-
-      router.navigate(".?a=1");
-      expect(createPath(router.state.location)).toBe("/base/path?a=1");
-      expect(router.state.matches[0].route.path).toBe("/path");
-
-      router.navigate("?b=2");
-      expect(createPath(router.state.location)).toBe("/base/path?b=2");
-      expect(router.state.matches[0].route.path).toBe("/path");
-
-      router.navigate("/path?c=3");
-      expect(createPath(router.state.location)).toBe("/base/path?c=3");
-      expect(router.state.matches[0].route.path).toBe("/path");
-    });
-  });
-
-  describe("view transitions", () => {
-    it("only enables view transitions when specified for the navigation", () => {
-      let t = setup({
-        routes: [{ path: "/" }, { path: "/a" }, { path: "/b" }],
-      });
-      let spy = jest.fn();
-      let unsubscribe = t.router.subscribe(spy);
-
-      // PUSH / -> /a - w/o transition
-      t.navigate("/a");
-      expect(spy).toHaveBeenLastCalledWith(
-        expect.objectContaining({
-          navigation: IDLE_NAVIGATION,
-          location: expect.objectContaining({ pathname: "/a" }),
-        }),
-        { unstable_viewTransitionOpts: undefined }
-      );
-
-      // PUSH /a -> /b - w/ transition
-      t.navigate("/b", { unstable_viewTransition: true });
-      expect(spy).toHaveBeenLastCalledWith(
-        expect.objectContaining({
-          navigation: IDLE_NAVIGATION,
-          location: expect.objectContaining({ pathname: "/b" }),
-        }),
-        {
-          unstable_viewTransitionOpts: {
-            currentLocation: expect.objectContaining({ pathname: "/a" }),
-            nextLocation: expect.objectContaining({ pathname: "/b" }),
-          },
-        }
-      );
-
-      // POP /b -> /a - w/ transition (cached from above)
-      t.navigate(-1);
-      expect(spy).toHaveBeenLastCalledWith(
-        expect.objectContaining({
-          navigation: IDLE_NAVIGATION,
-          location: expect.objectContaining({ pathname: "/a" }),
-        }),
-        {
-          unstable_viewTransitionOpts: {
-            // Args reversed on POP so same hooks apply
-            currentLocation: expect.objectContaining({ pathname: "/a" }),
-            nextLocation: expect.objectContaining({ pathname: "/b" }),
-          },
-        }
-      );
-
-      // POP /a -> / - No transition
-      t.navigate(-1);
-      expect(spy).toHaveBeenLastCalledWith(
-        expect.objectContaining({
-          navigation: IDLE_NAVIGATION,
-          location: expect.objectContaining({ pathname: "/" }),
-        }),
-        { unstable_viewTransitionOpts: undefined }
-      );
-
-      unsubscribe();
-      t.router.dispose();
     });
   });
 });
